+++ app/services/voice.py
@@ -1,256 +1,91 @@
 import asyncio
-import base64
-import io
-import inspect
 import json
-import math
 import os
-import queue
 import re
 import subprocess
-import tempfile
-import threading
-import time
 import unicodedata
-from datetime import datetime
 from typing import Union
-from urllib.parse import urlparse
-from xml.sax.saxutils import escape, unescape
+from xml.sax.saxutils import unescape
 
-import edge_tts
-import requests
+import edge_tts  # noqa: F401 (kept for `voice.edge_tts` back-compat access, e.g. in tests)
+import requests  # noqa: F401 (kept for `voice.requests` back-compat access, e.g. in tests)
 from edge_tts import SubMaker
 from loguru import logger
 from moviepy.video.tools import subtitles
 from moviepy.audio.io.AudioFileClip import AudioFileClip
-from openai import OpenAI
 
-from app.config import config
+from app.config import config  # noqa: F401 (kept for `voice.config` back-compat access, e.g. in tests)
 from app.utils import utils
 
-_DEFAULT_EDGE_TTS_TIMEOUT_SECONDS = 30.0
-_MIMO_DEFAULT_BASE_URL = "https://api.xiaomimimo.com/v1"
-_MIMO_DEFAULT_TTS_MODEL = "mimo-v2.5-tts"
-MINIMAX_TTS_GLOBAL_URL = "https://api.minimax.io/v1/t2a_v2"
-MINIMAX_TTS_CN_URL = "https://api.minimaxi.com/v1/t2a_v2"
-MINIMAX_TTS_DEFAULT_MODEL = "speech-2.8-hd"
-MINIMAX_TTS_DEFAULT_VOICE = "English_expressive_narrator"
-MINIMAX_TTS_MODELS = (
-    "speech-2.8-hd", "speech-2.8-turbo", "speech-2.6-hd", "speech-2.6-turbo",
-    "speech-02-hd", "speech-02-turbo", "speech-01-hd", "speech-01-turbo",
+# Provider implementations live in dedicated ``voice_<provider>.py`` modules;
+# this module stays a thin facade so ``voice.<name>`` keeps working for every
+# caller (WebUI, task pipeline, tests) exactly as before the split.
+from app.services.voice_common import (  # noqa: F401
+    convert_rate_to_percent,
+    ensure_file_path_exists,
+    ensure_legacy_submaker_fields,
+    is_azure_v2_voice,
+    mktimestamp,
+    parse_voice_name,
+    populate_legacy_submaker_with_full_text,
 )
-GEMINI_TTS_VOICES = (
-    ("Zephyr", "Bright"),
-    ("Puck", "Upbeat"),
-    ("Charon", "Informative"),
-    ("Kore", "Firm"),
-    ("Fenrir", "Excitable"),
-    ("Leda", "Youthful"),
-    ("Orus", "Firm"),
-    ("Aoede", "Breezy"),
-    ("Callirrhoe", "Easy-going"),
-    ("Autonoe", "Bright"),
-    ("Enceladus", "Breathy"),
-    ("Iapetus", "Clear"),
-    ("Umbriel", "Easy-going"),
-    ("Algieba", "Smooth"),
-    ("Despina", "Smooth"),
-    ("Erinome", "Clear"),
-    ("Algenib", "Gravelly"),
-    ("Rasalgethi", "Informative"),
-    ("Laomedeia", "Upbeat"),
-    ("Achernar", "Soft"),
-    ("Alnilam", "Firm"),
-    ("Schedar", "Even"),
-    ("Gacrux", "Mature"),
-    ("Pulcherrima", "Forward"),
-    ("Achird", "Friendly"),
-    ("Zubenelgenubi", "Casual"),
-    ("Vindemiatrix", "Gentle"),
-    ("Sadachbia", "Lively"),
-    ("Sadaltager", "Knowledgeable"),
-    ("Sulafat", "Warm"),
+from app.services.voice_edge import azure_tts_v1  # noqa: F401
+from app.services.voice_azure_v2 import azure_tts_v2  # noqa: F401
+from app.services.voice_siliconflow import (  # noqa: F401
+    get_siliconflow_voices,
+    is_siliconflow_voice,
+    siliconflow_tts,
 )
-_MINIMAX_TTS_MAX_AUDIO_HEX_CHARS = 100 * 1024 * 1024
+from app.services.voice_gemini import (  # noqa: F401
+    GEMINI_TTS_VOICES,
+    gemini_tts,
+    get_gemini_voices,
+    is_gemini_voice,
+    parse_gemini_voice_name,
+)
+from app.services.voice_mimo import (  # noqa: F401
+    get_mimo_voices,
+    is_mimo_voice,
+    mimo_tts,
+)
+from app.services.voice_minimax import (  # noqa: F401
+    MINIMAX_TTS_CN_URL,
+    MINIMAX_TTS_DEFAULT_MODEL,
+    MINIMAX_TTS_DEFAULT_VOICE,
+    MINIMAX_TTS_GLOBAL_URL,
+    MINIMAX_TTS_MODELS,
+    get_minimax_tts_api_key,
+    get_minimax_tts_endpoint,
+    get_minimax_voice_catalog,
+    get_minimax_voices,
+    is_minimax_voice,
+    minimax_tts,
+)
+from app.services.voice_elevenlabs import (  # noqa: F401
+    elevenlabs_tts,
+    get_elevenlabs_api_key,
+    get_elevenlabs_voices,
+    is_elevenlabs_voice,
+)
+from app.services.voice_chatterbox import (  # noqa: F401
+    chatterbox_tts,
+    get_chatterbox_voices,
+    is_chatterbox_voice,
+)
+from app.services.voice_fish_audio import (  # noqa: F401
+    FISH_AUDIO_DEFAULT_MODEL,
+    FISH_AUDIO_MODELS,
+    fish_audio_tts,
+    get_fish_audio_api_key,
+    get_fish_audio_voices,
+    is_fish_audio_voice,
+)
+
 NO_VOICE_NAME = "no-voice"
 # `none` 是 PR #981 里曾使用过的无配音标识。这里短期兼容这个值，避免
 # 已经手动调用过该分支的 API 用户升级后立即失效；WebUI 和新代码统一使用
 # 更明确的 `no-voice`。
 _NO_VOICE_ALIASES = {NO_VOICE_NAME, "none"}
-
-
-def _configure_pydub_ffmpeg(audio_segment_cls):
-    configured_ffmpeg = utils.get_ffmpeg_binary()
-    if configured_ffmpeg:
-        audio_segment_cls.converter = configured_ffmpeg
-
-
-def mktimestamp(time_unit: float) -> str:
-    """
-    将 edge_tts 使用的 100 纳秒时间单位转换为字幕时间戳。
-
-    edge_tts 7.x 不再导出旧版本里的 `mktimestamp`，但项目里旧字幕链路
-    还需要这个格式化函数来兼容 Azure v2、Gemini、SiliconFlow 这些
-    手工构造的字幕时间轴，因此这里内置一个等价实现。
-    """
-    hour = math.floor(time_unit / 10**7 / 3600)
-    minute = math.floor((time_unit / 10**7 / 60) % 60)
-    seconds = (time_unit / 10**7) % 60
-    return f"{hour:02d}:{minute:02d}:{seconds:06.3f}"
-
-
-def get_siliconflow_voices() -> list[str]:
-    """
-    获取硅基流动的声音列表
-
-    Returns:
-        声音列表，格式为 ["siliconflow:FunAudioLLM/CosyVoice2-0.5B:alex", ...]
-    """
-    # 硅基流动的声音列表和对应的性别（用于显示）
-    voices_with_gender = [
-        ("FunAudioLLM/CosyVoice2-0.5B", "alex", "Male"),
-        ("FunAudioLLM/CosyVoice2-0.5B", "anna", "Female"),
-        ("FunAudioLLM/CosyVoice2-0.5B", "bella", "Female"),
-        ("FunAudioLLM/CosyVoice2-0.5B", "benjamin", "Male"),
-        ("FunAudioLLM/CosyVoice2-0.5B", "charles", "Male"),
-        ("FunAudioLLM/CosyVoice2-0.5B", "claire", "Female"),
-        ("FunAudioLLM/CosyVoice2-0.5B", "david", "Male"),
-        ("FunAudioLLM/CosyVoice2-0.5B", "diana", "Female"),
-    ]
-
-    # 添加siliconflow:前缀，并格式化为显示名称
-    return [
-        f"siliconflow:{model}:{voice}-{gender}"
-        for model, voice, gender in voices_with_gender
-    ]
-
-
-def get_gemini_voices() -> list[str]:
-    """
-    获取 Gemini TTS 官方预置音色列表。
-
-    Google 没有为这些音色发布性别元数据，因此下拉框使用官方风格描述，
-    避免把推测的性别写进持久化 voice id。音色目录来源：
-    https://ai.google.dev/gemini-api/docs/speech-generation#voice-options
-
-    Returns:
-        声音列表，格式为 ["gemini:Zephyr-Bright", "gemini:Puck-Upbeat", ...]
-    """
-    return [f"gemini:{voice}-{style}" for voice, style in GEMINI_TTS_VOICES]
-
-
-def get_mimo_voices() -> list[str]:
-    """
-    获取 Xiaomi MiMo V2.5 TTS 的预置音色列表。
-
-    当前只接入官方文档里的 `mimo-v2.5-tts` 预置音色模式。音色设计
-    `mimo-v2.5-tts-voicedesign` 和音色复刻 `mimo-v2.5-tts-voiceclone`
-    需要额外的输入表单和素材上传流程，先不混入普通 TTS 下拉框，避免
-    用户误以为选择一个 voice id 就能完成所有高级能力。
-    """
-    voices_with_gender = [
-        ("mimo_default", "Female"),
-        ("冰糖", "Female"),
-        ("茉莉", "Female"),
-        ("苏打", "Male"),
-        ("白桦", "Male"),
-        ("Mia", "Female"),
-        ("Chloe", "Female"),
-        ("Milo", "Male"),
-        ("Dean", "Male"),
-    ]
-
-    return [f"mimo:{voice}-{gender}" for voice, gender in voices_with_gender]
-
-
-def get_minimax_voices(voice_id: str | None = None) -> list[str]:
-    """返回当前配置的 MiniMax 音色，供统一的 TTS 调度格式使用。"""
-    voice_id = str(
-        voice_id
-        or config.minimax_tts.get("voice_id", MINIMAX_TTS_DEFAULT_VOICE)
-        or MINIMAX_TTS_DEFAULT_VOICE
-    ).strip()
-    return [f"minimax:{voice_id}"]
-
-
-def get_elevenlabs_voices(api_key: str) -> list[str]:
-    if not api_key:
-        return []
-    try:
-        url = "https://api.elevenlabs.io/v2/voices"
-        params = {"is_favorite": "true", "page_size": 100}
-        headers = {"xi-api-key": api_key}
-        response = requests.get(url, params=params, headers=headers, timeout=10)
-        if response.status_code != 200:
-            logger.warning(
-                f"ElevenLabs voices fetch failed with status {response.status_code}: {response.text}"
-            )
-            return []
-        data = response.json()
-        voices = data.get("voices", [])
-        return [
-            f"elevenlabs:{v['voice_id']}:{v['name']}"
-            for v in voices
-            if v.get("voice_id") and v.get("name") and v.get("status") != "disabled"
-        ]
-    except Exception as e:
-        logger.warning(f"ElevenLabs voices fetch failed: {str(e)}")
-        return []
-
-
-def get_chatterbox_voices() -> list[str]:
-    """Return the configured Chatterbox voices.
-
-    Chatterbox is self-hosted, so there is no global voice catalog. Operators
-    list the voice names exposed by their server via ``[chatterbox] voices``
-    (a TOML array, or a comma-separated string). Each entry is normalised to
-    the ``chatterbox:<name>`` format used by the TTS dispatcher.
-    """
-    voices = config.chatterbox.get("voices", []) or []
-    if isinstance(voices, str):
-        voices = [v.strip() for v in voices.split(",") if v.strip()]
-    result = []
-    for v in voices:
-        v = str(v).strip()
-        if not v:
-            continue
-        result.append(v if v.startswith("chatterbox:") else f"chatterbox:{v}")
-    if not result:
-        # keep the dropdown usable even before any voice is configured
-        result = ["chatterbox:default-Female"]
-    return result
-
-
-def get_fish_audio_voices() -> list[str]:
-    """Return configured Fish Audio voices.
-
-    Each entry follows the format ``fish_audio:<reference_id>:<display_name>``.
-    When ``reference_id`` is "default", Fish Audio's built-in default voice is
-    used (no ``reference_id`` is sent in the API request).  Operators can list
-    additional public or cloned voices via ``[fish_audio] voices`` in the
-    config file.
-    """
-    result = [
-        "fish_audio:2324c907b9a94c64ab4afb941e5b3408:Clear Female-Female",
-        "fish_audio:7b6131ba75ba47c98a46c847db729ab6:Clear Male-Male",
-        "fish_audio:default:Default Voice",
-    ]
-    voices = config.fish_audio.get("voices", []) or []
-    if isinstance(voices, str):
-        voices = [v.strip() for v in voices.split(",") if v.strip()]
-    for entry in voices:
-        entry = str(entry).strip()
-        if not entry:
-            continue
-        if entry.startswith("fish_audio:"):
-            result.append(entry)
-        elif ":" in entry:
-            # "<reference_id>:<display_name>"
-            result.append(f"fish_audio:{entry}")
-        else:
-            # bare reference_id
-            result.append(f"fish_audio:{entry}:{entry}")
-    return result
 
 
 _AZURE_VOICES_DATA_FILE = os.path.join(
@@ -284,76 +119,6 @@ def get_all_azure_voices(filter_locals=None) -> list[str]:
     return voices
 
 
-def parse_voice_name(name: str):
-    # zh-CN-XiaoyiNeural-Female
-    # zh-CN-YunxiNeural-Male
-    # zh-CN-XiaoxiaoMultilingualNeural-V2-Female
-    name = name.replace("-Female", "").replace("-Male", "").strip()
-    return name
-
-
-def is_azure_v2_voice(voice_name: str):
-    voice_name = parse_voice_name(voice_name)
-    if voice_name.endswith("-V2"):
-        return voice_name.replace("-V2", "").strip()
-    return ""
-
-
-def is_siliconflow_voice(voice_name: str):
-    """检查是否是硅基流动的声音"""
-    return voice_name.startswith("siliconflow:")
-
-
-def is_gemini_voice(voice_name: str):
-    """检查是否是Gemini TTS的声音"""
-    return voice_name.startswith("gemini:")
-
-
-def parse_gemini_voice_name(voice_name: str | None) -> str:
-    """从新旧 Gemini 下拉框值中提取 Google API 使用的预置音色名称。"""
-    if not is_gemini_voice(voice_name or ""):
-        return ""
-    return (voice_name or "").split(":", 1)[1].split("-", 1)[0].strip()
-
-
-def is_mimo_voice(voice_name: str):
-    """检查是否是 Xiaomi MiMo TTS 的声音"""
-    return voice_name.startswith("mimo:")
-
-
-def is_minimax_voice(voice_name: str | None) -> bool:
-    return (voice_name or "").startswith("minimax:")
-
-
-def is_elevenlabs_voice(voice_name: str) -> bool:
-    return (voice_name or "").startswith("elevenlabs:")
-
-
-def get_elevenlabs_api_key() -> str:
-    """
-    读取 ElevenLabs TTS 使用的 API Key。
-
-    配置文件优先，环境变量仅作为未配置时的后备来源。WebUI 和配乐功能已经
-    支持 ``ELEVENLABS_API_KEY``，TTS 必须使用相同规则，否则仅通过容器环境
-    变量部署时，音色列表可正常加载，真正合成语音却会误报未配置 Key。
-    """
-    configured_key = str(config.elevenlabs.get("api_key", "") or "").strip()
-    return configured_key or os.getenv("ELEVENLABS_API_KEY", "").strip()
-
-
-def is_chatterbox_voice(voice_name: str) -> bool:
-    return (voice_name or "").startswith("chatterbox:")
-
-
-def is_fish_audio_voice(voice_name: str) -> bool:
-    return (voice_name or "").startswith("fish_audio:")
-
-
-def get_fish_audio_api_key() -> str:
-    configured_key = str(config.fish_audio.get("api_key", "") if hasattr(config, "fish_audio") and isinstance(config.fish_audio, dict) else "").strip()
-    return configured_key or os.getenv("FISH_API_KEY", "").strip()
-
-
 def is_no_voice(voice_name: str | None) -> bool:
     """
     判断用户是否明确选择了“无配音”模式。
@@ -382,7 +147,7 @@ def estimate_no_voice_duration(text: str) -> float:
     if not normalized_text:
         return 3.0
 
-    cjk_chars = len(re.findall(r"[\u4e00-\u9fff]", normalized_text))
+    cjk_chars = len(re.findall(r"[一-鿿]", normalized_text))
     words = len(re.findall(r"[A-Za-z0-9]+", normalized_text))
     ascii_word_chars = sum(len(word) for word in re.findall(r"[A-Za-z0-9]+", normalized_text))
     other_text_chars = 0
@@ -553,1340 +318,6 @@ def tts(
     return azure_tts_v1(text, voice_name, voice_rate, voice_file)
 
 
-def convert_rate_to_percent(rate: float) -> str:
-    # edge-tts requires a sign-prefixed percentage (e.g. "+0%", "-20%").
-    # Rounding can yield 0 for rates near but not equal to 1.0 (e.g. 1.004,
-    # 0.997); those must still be returned as "+0%", not the unsigned "0%"
-    # which edge-tts rejects with ValueError: Invalid rate '0%'.
-    # API 或批处理调用可能传入 0、0.0、None 或无法转换的空值；这些值不代表
-    # 合法语速，直接计算会变成 -100% 或抛异常。这里统一回退到正常语速，
-    # 避免生成极慢音频或让 TTS 流程在边界输入下失败。
-    try:
-        rate = float(rate)
-    except (TypeError, ValueError):
-        rate = 1.0
-    if rate <= 0:
-        rate = 1.0
-    percent = round((rate - 1.0) * 100)
-    if percent >= 0:
-        return f"+{percent}%"
-    return f"{percent}%"
-
-
-def ensure_file_path_exists(file_path: str) -> None:
-    """
-    确保输出文件所在目录一定存在。
-
-    这里单独做一层兜底，是因为 edge_tts 7.x 在真正发起网络请求之前，
-    就会先打开目标音频文件；如果目录不存在，会直接因为本地文件路径报错，
-    从而掩盖真正的 TTS 行为结果。
-    """
-    dir_path = os.path.dirname(file_path)
-    if dir_path:
-        os.makedirs(dir_path, exist_ok=True)
-
-
-def ensure_legacy_submaker_fields(sub_maker: SubMaker) -> SubMaker:
-    """
-    为项目里仍然沿用旧字幕结构的调用方补齐兼容字段。
-
-    edge_tts 7.x 的 `SubMaker` 主要暴露 `cues/get_srt()`，但项目里 Azure v2、
-    Gemini、SiliconFlow 这些路径仍然会直接读写 `subs/offset`。这里统一补齐，
-    避免升级 edge_tts 后这些非 edge 路径被连带破坏。
-    """
-    if not hasattr(sub_maker, "subs"):
-        sub_maker.subs = []
-    if not hasattr(sub_maker, "offset"):
-        sub_maker.offset = []
-    return sub_maker
-
-
-def populate_legacy_submaker_with_full_text(
-    sub_maker: SubMaker, text: str, audio_duration_seconds: float
-) -> SubMaker:
-    """
-    用整段文本填充项目历史沿用的 `subs/offset` 字幕结构。
-
-    背景：
-    1. edge_tts 7.x 的 `SubMaker` 不再提供旧版本里的 `create_sub()`；
-    2. 项目里 Gemini、SiliconFlow 等非 edge 路径依然需要返回一个
-       带 `subs/offset` 的对象，供后续统一计算音频时长和生成字幕；
-    3. 对于拿不到逐词边界的 TTS 服务，需要至少按脚本断句切成多个片段，
-       这样后续 `subtitle_provider=edge` 的聚合逻辑才能继续工作，而不是
-       因为整段文本无法和脚本断句逐行匹配而回退 Whisper。
-
-    Args:
-        sub_maker: 需要写入兼容字段的字幕对象
-        text: 原始脚本文本
-        audio_duration_seconds: 音频总时长，单位秒
-
-    Returns:
-        已填充兼容字幕数据的 SubMaker 对象
-    """
-    sub_maker = ensure_legacy_submaker_fields(sub_maker)
-
-    # 清空旧值，避免调用方重复复用对象时出现脏数据叠加。
-    sub_maker.subs = []
-    sub_maker.offset = []
-
-    normalized_text = (text or "").strip()
-    if not normalized_text:
-        return sub_maker
-
-    audio_duration_100ns = max(int(audio_duration_seconds * 10000000), 1)
-
-    # Gemini / SiliconFlow 这类路径拿不到逐词边界时，仍然尽量沿用项目
-    # 原来的“按标点断句 + 按字符数比例分配时长”的策略。这样既能让
-    # create_subtitle() 匹配脚本断句，也能避免再次回退 Whisper。
-    sentences = utils.split_string_by_punctuations(normalized_text)
-    if not sentences:
-        sentences = [normalized_text]
-
-    total_chars = sum(len(sentence) for sentence in sentences)
-    if total_chars <= 0:
-        sub_maker.subs.append(normalized_text)
-        sub_maker.offset.append((0, audio_duration_100ns))
-        return sub_maker
-
-    current_offset = 0
-    for index, sentence in enumerate(sentences):
-        cleaned_sentence = sentence.strip()
-        if not cleaned_sentence:
-            continue
-
-        # 前面的句子按字符数比例分配时长，最后一句兜底吃掉剩余时长，
-        # 避免整数取整导致总时长丢失或字幕结束时间短于音频。
-        if index == len(sentences) - 1:
-            sentence_end = audio_duration_100ns
-        else:
-            sentence_chars = len(cleaned_sentence)
-            sentence_duration = max(
-                int(audio_duration_100ns * (sentence_chars / total_chars)),
-                1,
-            )
-            sentence_end = min(current_offset + sentence_duration, audio_duration_100ns)
-
-        sub_maker.subs.append(cleaned_sentence)
-        sub_maker.offset.append((current_offset, sentence_end))
-        current_offset = sentence_end
-
-    return sub_maker
-
-
-def create_edge_tts_communicate(
-    text: str, voice_name: str, rate_str: str
-) -> edge_tts.Communicate:
-    """
-    按当前已安装的 edge_tts 版本构造 Communicate 对象。
-
-    背景：
-    1. 主线代码已经升级到 edge_tts 7.x，并使用 `boundary` 参数拿到更细的边界事件；
-    2. 但 Windows 便携包如果更新失败，现场环境可能仍然停留在旧版 edge_tts；
-    3. 旧版 `Communicate.__init__()` 不接受 `boundary`，会直接抛出
-       `unexpected keyword argument 'boundary'`，导致整个 TTS 链路失败。
-
-    因此这里先根据构造函数签名探测当前版本支持的参数，再决定是否传入
-    `boundary`，让同一份代码同时兼容旧版和新版依赖。
-    """
-    communicate_kwargs = {"rate": rate_str}
-    communicate_signature = inspect.signature(edge_tts.Communicate)
-
-    if "boundary" in communicate_signature.parameters:
-        communicate_kwargs["boundary"] = "WordBoundary"
-
-    return edge_tts.Communicate(text, voice_name, **communicate_kwargs)
-
-
-def get_edge_tts_timeout_seconds() -> Union[float, None]:
-    """
-    获取 Azure TTS V1 单次流式请求的超时时间。
-
-    背景：
-    Edge consumer TTS 在网络不通、服务端限流、voice 与文本语言不匹配等场景下，
-    可能长时间卡在 `stream_sync()` 内部，日志只停留在 `start`。这里提供一个
-    默认超时，避免 WebUI 任务长期无反馈。
-
-    使用方式：
-    - 默认 30 秒，覆盖常见短视频脚本的首包等待时间；
-    - 如用户处于慢网络或代理环境，可在 `config.toml` 里设置
-      `edge_tts_timeout = 60`；
-    - 设置为 0 或负数表示显式禁用超时，保留完全向后兼容。
-    """
-    raw_timeout = config.app.get(
-        "edge_tts_timeout", _DEFAULT_EDGE_TTS_TIMEOUT_SECONDS
-    )
-    try:
-        timeout_seconds = float(raw_timeout)
-    except (TypeError, ValueError):
-        logger.warning(
-            "invalid edge_tts_timeout: "
-            f"{raw_timeout}, fallback to {_DEFAULT_EDGE_TTS_TIMEOUT_SECONDS}s"
-        )
-        timeout_seconds = _DEFAULT_EDGE_TTS_TIMEOUT_SECONDS
-
-    if timeout_seconds <= 0:
-        return None
-
-    return timeout_seconds
-
-
-def _stream_edge_tts_sync_with_timeout(
-    communicate, on_chunk, timeout_seconds: float
-) -> None:
-    """
-    带总超时地消费 edge_tts 7.x 的同步流。
-
-    实现原因：
-    `stream_sync()` 本身是阻塞迭代器，网络层卡住时主线程无法及时恢复。
-    这里把阻塞迭代放到 daemon 线程中，主线程通过 Queue 获取 chunk，
-    到达超时时间后直接抛出 TimeoutError，让外层重试和错误日志继续工作。
-
-    注意：
-    daemon 线程只作为兜底保护使用，最多随 Azure TTS V1 的 3 次重试产生
-    少量残留线程；进程退出时会自动回收。相比 WebUI 任务永久卡住，这是
-    更可控的失败模式。
-    """
-    stream_queue = queue.Queue()
-    done_marker = object()
-
-    def _produce_chunks():
-        try:
-            for chunk in communicate.stream_sync():
-                stream_queue.put(("chunk", chunk))
-            stream_queue.put(("done", done_marker))
-        except Exception as e:
-            stream_queue.put(("error", e))
-
-    thread = threading.Thread(target=_produce_chunks, daemon=True)
-    thread.start()
-
-    deadline = time.monotonic() + timeout_seconds
-    while True:
-        remaining_seconds = deadline - time.monotonic()
-        if remaining_seconds <= 0:
-            raise TimeoutError(
-                f"edge_tts stream timed out after {timeout_seconds:g}s"
-            )
-
-        try:
-            item_type, payload = stream_queue.get(
-                timeout=min(0.5, remaining_seconds)
-            )
-        except queue.Empty:
-            continue
-
-        if item_type == "chunk":
-            on_chunk(payload)
-        elif item_type == "error":
-            raise payload
-        elif item_type == "done":
-            return
-
-
-def stream_edge_tts_chunks(
-    communicate, on_chunk, timeout_seconds: Union[float, None] = None
-) -> None:
-    """
-    统一消费 edge_tts 的同步流和旧版异步流。
-
-    edge_tts 7.x 提供 `stream_sync()`，可以在同步函数里直接迭代；
-    更早的版本通常只有异步 `stream()`。为了让 `azure_tts_v1()` 在
-    旧依赖残留场景下仍能继续工作，这里统一做一层流式兼容。
-
-    Args:
-        communicate: edge_tts.Communicate 实例
-        on_chunk: 每拿到一个事件块时执行的回调
-        timeout_seconds: 单次流式请求总超时；为 None 时不启用超时。
-    """
-    if hasattr(communicate, "stream_sync"):
-        if timeout_seconds:
-            _stream_edge_tts_sync_with_timeout(
-                communicate, on_chunk, timeout_seconds
-            )
-            return
-
-        for chunk in communicate.stream_sync():
-            on_chunk(chunk)
-        return
-
-    if not hasattr(communicate, "stream"):
-        raise AttributeError("edge_tts communicate object has no stream method")
-
-    async def _consume_async_stream():
-        async for chunk in communicate.stream():
-            on_chunk(chunk)
-
-    # 这里显式创建独立事件循环，而不是复用外部上下文，目的是避免
-    # 在同步调用栈里遇到“当前线程没有事件循环”或跨线程复用循环的问题。
-    loop = asyncio.new_event_loop()
-    try:
-        if timeout_seconds:
-            loop.run_until_complete(
-                asyncio.wait_for(_consume_async_stream(), timeout=timeout_seconds)
-            )
-        else:
-            loop.run_until_complete(_consume_async_stream())
-    finally:
-        loop.close()
-
-
-def azure_tts_v1(
-    text: str, voice_name: str, voice_rate: float, voice_file: str
-) -> Union[SubMaker, None]:
-    voice_name = parse_voice_name(voice_name)
-    text = text.strip()
-    rate_str = convert_rate_to_percent(voice_rate)
-    for i in range(3):
-        try:
-            logger.info(f"start, voice name: {voice_name}, try: {i + 1}")
-
-            # 这里同时兼容 edge_tts 7.x 和旧版便携包里可能残留的老依赖：
-            # 1. 新版支持 `boundary` + `stream_sync()`
-            # 2. 旧版不支持 `boundary`，且通常只暴露异步 `stream()`
-            ensure_file_path_exists(voice_file)
-            communicate = create_edge_tts_communicate(text, voice_name, rate_str)
-            sub_maker = edge_tts.SubMaker()
-            timeout_seconds = get_edge_tts_timeout_seconds()
-
-            with open(voice_file, "wb") as file:
-                def _handle_chunk(chunk):
-                    chunk_type = chunk["type"]
-                    if chunk_type == "audio":
-                        file.write(chunk["data"])
-                    elif chunk_type in ["WordBoundary", "SentenceBoundary"]:
-                        # 无论来自 7.x 的同步流，还是旧版异步流，只要事件结构
-                        # 里仍有边界信息，就统一喂给 SubMaker，保证后续字幕链路
-                        # 仍然走项目现有逻辑。
-                        sub_maker.feed(chunk)
-
-                stream_edge_tts_chunks(
-                    communicate, _handle_chunk, timeout_seconds=timeout_seconds
-                )
-
-            if not sub_maker.get_srt():
-                logger.warning("failed, sub_maker.get_srt() is empty")
-                continue
-
-            logger.info(f"completed, output file: {voice_file}")
-            return sub_maker
-        except Exception as e:
-            logger.error(f"failed, error: {str(e)}")
-            # TTS 流式写入如果在首包前超时或网络异常，会留下 0 字节音频文件。
-            # 这种文件既不可播放，也可能误导后续排查，因此失败后只清理空文件；
-            # 如果已经写入了部分数据，则保留现场文件，便于分析服务端返回内容。
-            if os.path.exists(voice_file) and os.path.getsize(voice_file) == 0:
-                try:
-                    os.remove(voice_file)
-                except Exception as remove_error:
-                    logger.warning(
-                        "failed to remove empty tts file: "
-                        f"{voice_file}, error: {str(remove_error)}"
-                    )
-    return None
-
-
-def siliconflow_tts(
-    text: str,
-    model: str,
-    voice: str,
-    voice_rate: float,
-    voice_file: str,
-    voice_volume: float = 1.0,
-) -> Union[SubMaker, None]:
-    """
-    使用硅基流动的API生成语音
-
-    Args:
-        text: 要转换为语音的文本
-        model: 模型名称，如 "FunAudioLLM/CosyVoice2-0.5B"
-        voice: 声音名称，如 "FunAudioLLM/CosyVoice2-0.5B:alex"
-        voice_rate: 语音速度，范围[0.25, 4.0]
-        voice_file: 输出的音频文件路径
-        voice_volume: 语音音量，范围[0.6, 5.0]，需要转换为硅基流动的增益范围[-10, 10]
-
-    Returns:
-        SubMaker对象或None
-    """
-    text = text.strip()
-    api_key = config.siliconflow.get("api_key", "")
-
-    if not api_key:
-        logger.error("SiliconFlow API key is not set")
-        return None
-
-    # 将voice_volume转换为硅基流动的增益范围
-    # 默认voice_volume为1.0，对应gain为0
-    gain = voice_volume - 1.0
-    # 确保gain在[-10, 10]范围内
-    gain = max(-10, min(10, gain))
-
-    url = "https://api.siliconflow.cn/v1/audio/speech"
-
-    payload = {
-        "model": model,
-        "input": text,
-        "voice": voice,
-        "response_format": "mp3",
-        "sample_rate": 32000,
-        "stream": False,
-        "speed": voice_rate,
-        "gain": gain,
-    }
-
-    headers = {"Authorization": f"Bearer {api_key}", "Content-Type": "application/json"}
-
-    for i in range(3):  # 尝试3次
-        try:
-            logger.info(
-                f"start siliconflow tts, model: {model}, voice: {voice}, try: {i + 1}"
-            )
-
-            response = requests.post(url, json=payload, headers=headers)
-
-            if response.status_code == 200:
-                # 保存音频文件
-                with open(voice_file, "wb") as f:
-                    f.write(response.content)
-
-                sub_maker = ensure_legacy_submaker_fields(SubMaker())
-
-                try:
-                    audio_clip = AudioFileClip(voice_file)
-                    try:
-                        audio_duration = audio_clip.duration
-                    finally:
-                        audio_clip.close()
-                except Exception as e:
-                    logger.warning(f"Failed to read audio duration: {str(e)}")
-                    audio_duration = 10.0
-
-                logger.success(f"siliconflow tts succeeded: {voice_file}")
-                return populate_legacy_submaker_with_full_text(
-                    sub_maker=sub_maker,
-                    text=text,
-                    audio_duration_seconds=audio_duration,
-                )
-            else:
-                logger.error(
-                    f"siliconflow tts failed with status code {response.status_code}: {response.text}"
-                )
-        except Exception as e:
-            logger.error(f"siliconflow tts failed: {str(e)}")
-
-    return None
-
-
-def _build_azure_v2_ssml(text: str, voice_name: str, voice_rate: float) -> str:
-    """构造 Azure Speech V2 使用的 SSML，并安全规范化语速参数。"""
-    try:
-        normalized_rate = float(voice_rate)
-    except (TypeError, ValueError):
-        normalized_rate = 1.0
-    normalized_rate = max(0.25, min(4.0, normalized_rate))
-
-    voice_locale_parts = voice_name.split("-", 2)
-    voice_locale = (
-        "-".join(voice_locale_parts[:2])
-        if len(voice_locale_parts) >= 2
-        else "en-US"
-    )
-    escaped_text = escape(text)
-    escaped_voice_name = escape(voice_name, {'"': "&quot;"})
-    return (
-        '<speak version="1.0" xmlns="http://www.w3.org/2001/10/synthesis" '
-        f'xml:lang="{voice_locale}">'
-        f'<voice name="{escaped_voice_name}">'
-        f'<prosody rate="{normalized_rate:g}">{escaped_text}</prosody>'
-        "</voice></speak>"
-    )
-
-
-def azure_tts_v2(
-    text: str,
-    voice_name: str,
-    voice_file: str,
-    voice_rate: float = 1.0,
-) -> Union[SubMaker, None]:
-    voice_name = is_azure_v2_voice(voice_name)
-    if not voice_name:
-        logger.error(f"invalid voice name: {voice_name}")
-        raise ValueError(f"invalid voice name: {voice_name}")
-    text = text.strip()
-    ssml = _build_azure_v2_ssml(text, voice_name, voice_rate)
-
-    def _format_duration_to_offset(duration) -> int:
-        if isinstance(duration, str):
-            time_obj = datetime.strptime(duration, "%H:%M:%S.%f")
-            milliseconds = (
-                (time_obj.hour * 3600000)
-                + (time_obj.minute * 60000)
-                + (time_obj.second * 1000)
-                + (time_obj.microsecond // 1000)
-            )
-            return milliseconds * 10000
-
-        if isinstance(duration, int):
-            return duration
-
-        return 0
-
-    for i in range(3):
-        try:
-            logger.info(
-                f"start, voice name: {voice_name}, rate: {voice_rate}, try: {i + 1}"
-            )
-
-            import azure.cognitiveservices.speech as speechsdk
-
-            sub_maker = ensure_legacy_submaker_fields(SubMaker())
-
-            def speech_synthesizer_word_boundary_cb(evt: speechsdk.SessionEventArgs):
-                # print('WordBoundary event:')
-                # print('\tBoundaryType: {}'.format(evt.boundary_type))
-                # print('\tAudioOffset: {}ms'.format((evt.audio_offset + 5000)))
-                # print('\tDuration: {}'.format(evt.duration))
-                # print('\tText: {}'.format(evt.text))
-                # print('\tTextOffset: {}'.format(evt.text_offset))
-                # print('\tWordLength: {}'.format(evt.word_length))
-
-                duration = _format_duration_to_offset(str(evt.duration))
-                offset = _format_duration_to_offset(evt.audio_offset)
-                sub_maker.subs.append(evt.text)
-                sub_maker.offset.append((offset, offset + duration))
-
-            # Creates an instance of a speech config with specified subscription key and service region.
-            speech_key = config.azure.get("speech_key", "")
-            service_region = config.azure.get("speech_region", "")
-            if not speech_key or not service_region:
-                logger.error("Azure speech key or region is not set")
-                return None
-
-            audio_config = speechsdk.audio.AudioOutputConfig(
-                filename=voice_file, use_default_speaker=True
-            )
-            speech_config = speechsdk.SpeechConfig(
-                subscription=speech_key, region=service_region
-            )
-            speech_config.speech_synthesis_voice_name = voice_name
-            # speech_config.set_property(property_id=speechsdk.PropertyId.SpeechServiceResponse_RequestSentenceBoundary,
-            #                            value='true')
-            speech_config.set_property(
-                property_id=speechsdk.PropertyId.SpeechServiceResponse_RequestWordBoundary,
-                value="true",
-            )
-
-            speech_config.set_speech_synthesis_output_format(
-                speechsdk.SpeechSynthesisOutputFormat.Audio48Khz192KBitRateMonoMp3
-            )
-            speech_synthesizer = speechsdk.SpeechSynthesizer(
-                audio_config=audio_config, speech_config=speech_config
-            )
-            speech_synthesizer.synthesis_word_boundary.connect(
-                speech_synthesizer_word_boundary_cb
-            )
-
-            # speak_text_async() 不支持语速参数。使用 SSML prosody 后，试听和
-            # 正式生成都会按 WebUI/API 传入的 voice_rate 调整语速。
-            result = speech_synthesizer.speak_ssml_async(ssml).get()
-            if result.reason == speechsdk.ResultReason.SynthesizingAudioCompleted:
-                logger.success(f"azure v2 speech synthesis succeeded: {voice_file}")
-                return sub_maker
-            elif result.reason == speechsdk.ResultReason.Canceled:
-                cancellation_details = result.cancellation_details
-                logger.error(
-                    f"azure v2 speech synthesis canceled: {cancellation_details.reason}"
-                )
-                if cancellation_details.reason == speechsdk.CancellationReason.Error:
-                    logger.error(
-                        f"azure v2 speech synthesis error: {cancellation_details.error_details}"
-                    )
-            logger.info(f"completed, output file: {voice_file}")
-        except Exception as e:
-            logger.error(f"failed, error: {str(e)}")
-    return None
-
-
-def gemini_tts(
-    text: str,
-    voice_name: str,
-    voice_rate: float,
-    voice_file: str,
-    voice_volume: float = 1.0,
-) -> Union[SubMaker, None]:
-    """
-    使用Google Gemini TTS生成语音
-    
-    Args:
-        text: 要转换的文本
-        voice_name: 语音名称，如 "Zephyr", "Puck" 等
-        voice_rate: 语音速率（当前未使用）
-        voice_file: 输出音频文件路径
-        voice_volume: 音频音量（当前未使用）
-        
-    Returns:
-        SubMaker对象或None
-    """
-    import base64
-    import io
-    from pydub import AudioSegment
-    from google import genai
-    from google.genai import types
-    _configure_pydub_ffmpeg(AudioSegment)
-    
-    try:
-        api_key = config.app.get("gemini_api_key", "")
-        if not api_key:
-            logger.error("Gemini API key is not set")
-            return None
-
-        logger.info(f"start, voice name: {voice_name}, try: 1")
-
-        generation_config = types.GenerateContentConfig(
-            response_modalities=["AUDIO"],
-            speech_config=types.SpeechConfig(
-                voice_config=types.VoiceConfig(
-                    prebuilt_voice_config=types.PrebuiltVoiceConfig(
-                        voice_name=voice_name
-                    )
-                )
-            ),
-        )
-
-        # google-genai 使用统一 Client 调用文本和 TTS 模型。上下文管理器确保
-        # 请求结束后释放 HTTP 连接，同时保留原有 PCM 转码和字幕时间轴逻辑。
-        with genai.Client(api_key=api_key) as client:
-            response = client.models.generate_content(
-                model="gemini-2.5-flash-preview-tts",
-                contents=text,
-                config=generation_config,
-            )
-
-        # 检查响应
-        if not response.candidates or not response.candidates[0].content:
-            logger.error("No audio content received from Gemini TTS")
-            return None
-            
-        # 获取音频数据
-        audio_data = None
-        for part in response.candidates[0].content.parts:
-            if hasattr(part, 'inline_data') and part.inline_data:
-                audio_data = part.inline_data.data
-                break
-                
-        if not audio_data:
-            logger.error("No audio data found in response")
-            return None
-            
-        # 音频数据已经是原始字节，不需要base64解码
-        if isinstance(audio_data, str):
-            # 如果是字符串，则需要base64解码
-            audio_bytes = base64.b64decode(audio_data)
-        else:
-            # 如果已经是字节，直接使用
-            audio_bytes = audio_data
-        
-        # 尝试不同的音频格式 - Gemini可能返回不同的格式
-        audio_segment = None
-        
-        # Gemini返回Linear PCM格式，按照文档参数解析
-        try:
-            audio_segment = AudioSegment.from_file(
-                io.BytesIO(audio_bytes), 
-                format="raw",
-                frame_rate=24000,  # Gemini TTS默认采样率
-                channels=1,        # 单声道
-                sample_width=2     # 16-bit
-            )
-        except Exception as e:
-            logger.error(f"Failed to load PCM audio: {e}")
-            return None
-        
-        # API、CLI 或测试可以直接把尚不存在的嵌套目录作为输出位置。这里在
-        # 真正写文件前统一创建父目录，避免一次成功的 Gemini 请求最后因为
-        # 本地路径不存在而丢失结果，也让该 provider 与其他 TTS 实现行为一致。
-        ensure_file_path_exists(voice_file)
-
-        # pydub 会返回打开的输出文件对象。批量生成时若不主动关闭，文件描述符
-        # 会持续累积，并在 Windows 上增加后续覆盖或删除音频文件失败的概率。
-        exported_audio = audio_segment.export(voice_file, format="mp3")
-        exported_audio.close()
-        
-        logger.info(f"completed, output file: {voice_file}")
-        
-        # Gemini 拿不到 edge_tts 那种逐词边界事件，因此这里退回到
-        # 项目原有的 `subs/offset` 兼容结构，至少保证后续字幕与时长
-        # 计算链路可继续工作。
-        sub_maker = ensure_legacy_submaker_fields(SubMaker())
-        audio_duration = len(audio_segment) / 1000.0  # 转换为秒
-        return populate_legacy_submaker_with_full_text(
-            sub_maker=sub_maker,
-            text=text,
-            audio_duration_seconds=audio_duration,
-        )
-        
-    except ImportError as e:
-        logger.error(f"Missing required package for Gemini TTS: {str(e)}. Please install: pip install pydub")
-        return None
-    except Exception as e:
-        logger.error(f"Gemini TTS failed, error: {str(e)}")
-        return None
-
-
-def mimo_tts(
-    text: str,
-    voice_name: str,
-    voice_rate: float,
-    voice_file: str,
-    voice_volume: float = 1.0,
-) -> Union[SubMaker, None]:
-    """
-    使用 Xiaomi MiMo V2.5 TTS 生成语音。
-
-    官方接口兼容 OpenAI Chat Completions，但 TTS 有两个关键差异：
-    1. 待合成文本必须放在 `assistant` 消息里；
-    2. 音频以 `message.audio.data` 的 base64 字符串返回。
-
-    MiMo 当前没有返回逐词时间轴，因此这里复用项目已有的 legacy
-    SubMaker 兜底方案：根据最终音频时长和脚本文本断句生成字幕时间轴。
-    """
-    from pydub import AudioSegment
-
-    text = (text or "").strip()
-    if not text:
-        logger.error("MiMo TTS text is empty")
-        return None
-
-    api_key = config.app.get("mimo_api_key", "")
-    if not api_key:
-        logger.error("MiMo API key is not set")
-        return None
-
-    base_url = config.app.get("mimo_base_url", "") or _MIMO_DEFAULT_BASE_URL
-    model_name = config.app.get("mimo_tts_model_name", "") or _MIMO_DEFAULT_TTS_MODEL
-    style_prompt = config.app.get(
-        "mimo_tts_style_prompt",
-        "请用自然、清晰、适合短视频旁白的语气朗读。",
-    )
-
-    _configure_pydub_ffmpeg(AudioSegment)
-
-    for i in range(3):
-        try:
-            logger.info(
-                f"start mimo tts, model: {model_name}, voice: {voice_name}, try: {i + 1}"
-            )
-            ensure_file_path_exists(voice_file)
-
-            client = OpenAI(api_key=api_key, base_url=base_url)
-            completion = client.chat.completions.create(
-                model=model_name,
-                messages=[
-                    {"role": "user", "content": style_prompt},
-                    {"role": "assistant", "content": text},
-                ],
-                audio={
-                    "format": "wav",
-                    "voice": voice_name,
-                },
-            )
-
-            if not completion or not getattr(completion, "choices", None):
-                raise ValueError("MiMo TTS returned empty response")
-
-            message = completion.choices[0].message
-            audio = getattr(message, "audio", None)
-            audio_data = None
-            if isinstance(audio, dict):
-                audio_data = audio.get("data")
-            elif audio is not None:
-                audio_data = getattr(audio, "data", None)
-
-            if not audio_data:
-                raise ValueError("MiMo TTS returned empty audio data")
-
-            audio_bytes = base64.b64decode(audio_data)
-            audio_segment = AudioSegment.from_file(io.BytesIO(audio_bytes), format="wav")
-
-            output_format = utils.parse_extension(voice_file) or "mp3"
-            if output_format == "wav":
-                with open(voice_file, "wb") as f:
-                    f.write(audio_bytes)
-            else:
-                audio_segment.export(voice_file, format=output_format)
-
-            audio_duration = len(audio_segment) / 1000.0
-            sub_maker = ensure_legacy_submaker_fields(SubMaker())
-            logger.success(f"mimo tts succeeded: {voice_file}")
-            logger.debug(
-                "mimo subtitle timeline generated, "
-                f"duration: {audio_duration:.3f}s, output_format: {output_format}"
-            )
-            return populate_legacy_submaker_with_full_text(
-                sub_maker=sub_maker,
-                text=text,
-                audio_duration_seconds=audio_duration,
-            )
-        except Exception as e:
-            logger.error(f"mimo tts failed: {str(e)}")
-
-    return None
-
-
-def _resolve_minimax_tts_url(configured_url: str) -> str:
-    configured_url = (configured_url or "").strip().rstrip("/")
-    if not configured_url:
-        return MINIMAX_TTS_GLOBAL_URL
-    if configured_url in {MINIMAX_TTS_GLOBAL_URL, MINIMAX_TTS_CN_URL}:
-        return configured_url
-    if configured_url.endswith("/v1"):
-        return f"{configured_url}/t2a_v2"
-    return configured_url
-
-
-def get_minimax_tts_api_key() -> str:
-    """返回 MiniMax TTS 的有效密钥，专用配置优先于 LLM 共享配置。"""
-    return str(
-        config.minimax_tts.get("api_key", "")
-        or config.app.get("minimax_api_key", "")
-        or os.getenv("MINIMAX_API_KEY", "")
-        or ""
-    ).strip()
-
-
-def _infer_minimax_tts_url(base_url: str) -> str:
-    """根据 MiniMax LLM 地址推断同区域的 TTS 地址，无法识别时返回空值。"""
-    normalized_url = str(base_url or "").strip()
-    if not normalized_url:
-        return ""
-
-    parse_target = normalized_url if "://" in normalized_url else f"//{normalized_url}"
-    host = (urlparse(parse_target).hostname or "").lower()
-    if host == "minimaxi.com" or host.endswith(".minimaxi.com"):
-        return MINIMAX_TTS_CN_URL
-    if host == "minimax.io" or host.endswith(".minimax.io"):
-        return MINIMAX_TTS_GLOBAL_URL
-    return ""
-
-
-def get_minimax_tts_endpoint() -> str:
-    """
-    返回与当前有效密钥匹配的 MiniMax TTS 地址。
-
-    独立配置 TTS Key 时尊重用户选择的 TTS 地址；复用 MiniMax LLM Key 时，
-    优先跟随 LLM Base URL 的区域，避免中国站 Key 被发送到国际站而返回 401。
-    """
-    dedicated_key = str(config.minimax_tts.get("api_key", "") or "").strip()
-    if not dedicated_key:
-        inferred_url = _infer_minimax_tts_url(config.app.get("minimax_base_url", ""))
-        if inferred_url:
-            return inferred_url
-    return _resolve_minimax_tts_url(config.minimax_tts.get("base_url", ""))
-
-
-def get_minimax_voice_catalog(
-    api_key: str = "",
-    endpoint: str = "",
-    voice_type: str = "all",
-) -> list[dict[str, str]]:
-    """
-    查询当前 MiniMax 账号可用的系统、克隆和生成音色。
-
-    返回值统一为 voice_id、voice_name、voice_type 三个字段，调用方无需了解
-    MiniMax 按音色来源拆分数组的响应结构。查询失败时抛出异常，让 WebUI、
-    API 或 CLI 可以按各自交互方式展示明确错误，而不是静默返回空列表。
-    """
-    if voice_type not in {"system", "voice_cloning", "voice_generation", "all"}:
-        raise ValueError(f"Unsupported MiniMax voice type: {voice_type}")
-
-    effective_api_key = str(api_key or get_minimax_tts_api_key()).strip()
-    if not effective_api_key:
-        raise ValueError("MiniMax TTS API key is not set")
-
-    tts_endpoint = (
-        _resolve_minimax_tts_url(endpoint)
-        if endpoint
-        else get_minimax_tts_endpoint()
-    )
-    voice_endpoint = (
-        f"{tts_endpoint[:-len('/t2a_v2')]}/get_voice"
-        if tts_endpoint.endswith("/t2a_v2")
-        else f"{tts_endpoint.rstrip('/')}/get_voice"
-    )
-    response = requests.post(
-        voice_endpoint,
-        json={"voice_type": voice_type},
-        headers={
-            "Authorization": f"Bearer {effective_api_key}",
-            "Content-Type": "application/json",
-        },
-        timeout=30,
-    )
-    if response.status_code != 200:
-        raise RuntimeError(
-            f"MiniMax get_voice failed with status {response.status_code}: "
-            f"{response.text[:200]}"
-        )
-
-    try:
-        body = response.json()
-    except ValueError as exc:
-        raise RuntimeError("MiniMax get_voice returned invalid JSON") from exc
-
-    base_resp = body.get("base_resp") or {}
-    if base_resp.get("status_code") not in {0, "0"}:
-        status_message = str(base_resp.get("status_msg") or "unknown error")
-        raise RuntimeError(f"MiniMax get_voice failed: {status_message}")
-
-    catalog = []
-    seen_voice_ids = set()
-    response_groups = (
-        ("system", "system_voice"),
-        ("voice_cloning", "voice_cloning"),
-        ("voice_generation", "voice_generation"),
-    )
-    for normalized_type, response_key in response_groups:
-        for item in body.get(response_key) or []:
-            voice_id = str(item.get("voice_id") or "").strip()
-            if not voice_id or voice_id in seen_voice_ids:
-                continue
-            seen_voice_ids.add(voice_id)
-            catalog.append(
-                {
-                    "voice_id": voice_id,
-                    "voice_name": str(item.get("voice_name") or voice_id).strip(),
-                    "voice_type": normalized_type,
-                }
-            )
-
-    logger.info(f"loaded MiniMax voices: count={len(catalog)}, type={voice_type}")
-    return catalog
-
-
-def _write_validated_minimax_audio(audio_bytes: bytes, voice_file: str) -> float:
-    """
-    将 MiniMax 音频原子写入目标路径，并返回时长。
-
-    远端返回成功状态不代表音频一定完整。先在同目录临时文件中验证，再使用
-    os.replace 原子替换，可以避免解码失败或 MoviePy 无法读取时留下半成品。
-    """
-    ensure_file_path_exists(voice_file)
-    output_dir = os.path.dirname(os.path.abspath(voice_file))
-    output_suffix = os.path.splitext(voice_file)[1] or ".mp3"
-    temp_fd, temp_path = tempfile.mkstemp(
-        prefix=".minimax-tts-", suffix=output_suffix, dir=output_dir
-    )
-    os.close(temp_fd)
-
-    try:
-        with open(temp_path, "wb") as output:
-            output.write(audio_bytes)
-
-        audio_clip = AudioFileClip(temp_path)
-        try:
-            audio_duration = float(audio_clip.duration)
-        finally:
-            audio_clip.close()
-
-        if not math.isfinite(audio_duration) or audio_duration <= 0:
-            raise ValueError("MiniMax TTS returned audio with an invalid duration")
-
-        os.replace(temp_path, voice_file)
-        return audio_duration
-    finally:
-        if os.path.exists(temp_path):
-            os.remove(temp_path)
-
-
-def minimax_tts(text: str, voice_id: str, voice_rate: float, voice_file: str, voice_volume: float = 1.0) -> Union[SubMaker, None]:
-    """Generate speech with the synchronous MiniMax T2A HTTP API."""
-    text, voice_id = (text or "").strip(), (voice_id or "").strip()
-    if not text or not voice_id:
-        logger.error("MiniMax TTS requires text and a voice ID")
-        return None
-    settings = config.minimax_tts
-    api_key = get_minimax_tts_api_key()
-    if not api_key:
-        logger.error("MiniMax TTS API key is not set")
-        return None
-    url = get_minimax_tts_endpoint()
-    model = str(settings.get("model_id", MINIMAX_TTS_DEFAULT_MODEL) or MINIMAX_TTS_DEFAULT_MODEL).strip()
-    if model not in MINIMAX_TTS_MODELS:
-        logger.error(f"Unsupported MiniMax TTS model: {model}")
-        return None
-    try:
-        speed = max(0.5, min(2.0, float(voice_rate or 1.0)))
-        volume = max(0.0, min(10.0, float(voice_volume or 1.0)))
-        pitch = max(-12, min(12, int(settings.get("pitch", 0) or 0)))
-        sample_rate = int(settings.get("sample_rate", 32000) or 32000)
-        bitrate = int(settings.get("bitrate", 128000) or 128000)
-        channel = int(settings.get("channel", 1) or 1)
-    except (TypeError, ValueError) as exc:
-        logger.error(f"Invalid MiniMax TTS audio setting: {str(exc)}")
-        return None
-    audio_format = str(settings.get("audio_format", "mp3") or "mp3").strip()
-    if audio_format not in {"mp3", "wav", "flac", "pcm"}:
-        logger.error(f"Unsupported MiniMax TTS audio format: {audio_format}")
-        return None
-    payload = {
-        "model": model, "text": text, "stream": False, "language_boost": "auto", "output_format": "hex",
-        "voice_setting": {"voice_id": voice_id, "speed": speed, "vol": volume, "pitch": pitch},
-        "audio_setting": {"sample_rate": sample_rate, "bitrate": bitrate, "format": audio_format, "channel": channel},
-    }
-    headers = {"Authorization": f"Bearer {api_key}", "Content-Type": "application/json"}
-    for attempt in range(3):
-        try:
-            logger.info(f"start MiniMax TTS, model: {model}, voice: {voice_id}, try: {attempt + 1}")
-            response = requests.post(url, json=payload, headers=headers, timeout=120)
-            if response.status_code != 200:
-                logger.error(f"MiniMax TTS failed with status {response.status_code}: {response.text[:200]}")
-                continue
-            body = response.json()
-            data = body.get("data") or {}
-            base_resp = body.get("base_resp") or {}
-            if base_resp.get("status_code") != 0 or data.get("status") != 2:
-                logger.error(f"MiniMax TTS returned an unsuccessful response: status_code={base_resp.get('status_code')}, audio_status={data.get('status')}")
-                continue
-            audio_hex = data.get("audio")
-            if not isinstance(audio_hex, str) or not audio_hex:
-                logger.error("MiniMax TTS returned empty audio data")
-                continue
-            if len(audio_hex) > _MINIMAX_TTS_MAX_AUDIO_HEX_CHARS:
-                logger.error("MiniMax TTS returned audio data exceeding the supported size")
-                continue
-            audio_duration = _write_validated_minimax_audio(bytes.fromhex(audio_hex), voice_file)
-            logger.success(f"MiniMax TTS succeeded: {voice_file}")
-            return populate_legacy_submaker_with_full_text(
-                ensure_legacy_submaker_fields(SubMaker()), text, audio_duration
-            )
-        except (OSError, ValueError, requests.RequestException) as exc:
-            logger.error(f"MiniMax TTS failed: {str(exc)}")
-    return None
-
-
-def elevenlabs_tts(
-    text: str,
-    voice_id: str,
-    voice_file: str,
-    voice_rate: float = 1.0,
-    voice_volume: float = 1.0,
-    model_id: str = "",
-) -> Union[SubMaker, None]:
-    text = (text or "").strip()
-    if not text:
-        logger.error("ElevenLabs TTS text is empty")
-        return None
-
-    api_key = get_elevenlabs_api_key()
-    if not api_key:
-        logger.error("ElevenLabs API key is not set")
-        return None
-
-    if not model_id:
-        model_id = config.elevenlabs.get("model_id", "eleven_multilingual_v2")
-
-    url = f"https://api.elevenlabs.io/v1/text-to-speech/{voice_id}"
-    headers = {
-        "xi-api-key": api_key,
-        "Content-Type": "application/json",
-    }
-    payload = {
-        "text": text,
-        "model_id": model_id,
-        "voice_settings": {
-            "stability": 0.5,
-            "similarity_boost": 0.75,
-            "style": 0.0,
-            "use_speaker_boost": True,
-        },
-    }
-
-    # Errors where retrying will never help (auth/access/validation failures).
-    _NON_RETRYABLE_CODES = {401, 403, 422}
-    _NON_RETRYABLE_STATUSES = {"voice_disabled", "voice_access_denied", "unauthorized"}
-
-    for i in range(3):
-        try:
-            logger.info(f"start elevenlabs tts, voice_id: {voice_id}, try: {i + 1}")
-            ensure_file_path_exists(voice_file)
-
-            response = requests.post(url, json=payload, headers=headers, timeout=60)
-            if response.status_code != 200:
-                error_status = ""
-                try:
-                    detail = response.json().get("detail", {})
-                    if isinstance(detail, dict):
-                        error_status = detail.get("status", "")
-                except Exception:
-                    pass
-
-                if response.status_code in _NON_RETRYABLE_CODES or error_status in _NON_RETRYABLE_STATUSES:
-                    logger.error(
-                        f"ElevenLabs TTS failed (non-retryable) — voice_id: {voice_id}, "
-                        f"status: {response.status_code}, error: {error_status or response.text[:200]}. "
-                        "Please select a different ElevenLabs voice."
-                    )
-                    return None
-
-                logger.error(
-                    f"elevenlabs tts failed with status {response.status_code}: {response.text[:200]}"
-                )
-                continue
-
-            with open(voice_file, "wb") as f:
-                f.write(response.content)
-
-            audio_clip = AudioFileClip(voice_file)
-            try:
-                audio_duration = audio_clip.duration
-            finally:
-                audio_clip.close()
-
-            sub_maker = ensure_legacy_submaker_fields(SubMaker())
-            logger.success(f"elevenlabs tts succeeded: {voice_file}")
-            return populate_legacy_submaker_with_full_text(
-                sub_maker=sub_maker,
-                text=text,
-                audio_duration_seconds=audio_duration,
-            )
-        except Exception as e:
-            logger.error(f"elevenlabs tts failed: {str(e)}")
-
-    return None
-
-
-def chatterbox_tts(
-    text: str,
-    voice: str,
-    voice_file: str,
-    voice_rate: float = 1.0,
-    voice_volume: float = 1.0,
-    model_id: str = "",
-) -> Union[SubMaker, None]:
-    """Generate speech with a self-hosted Chatterbox TTS server.
-
-    Chatterbox (Resemble AI, MIT) is an open-source, locally hosted TTS model
-    with zero-shot voice cloning — a self-hostable alternative to ElevenLabs.
-    This talks to an OpenAI-compatible ``/audio/speech`` endpoint, so it works
-    with the common community servers (e.g. devnen/Chatterbox-TTS-Server,
-    travisvn/chatterbox-tts-api). Configure ``[chatterbox] base_url`` (and an
-    optional ``api_key``).
-
-    Like ElevenLabs, Chatterbox does not return word-level timestamps, so the
-    subtitle path falls back to the full-text SubMaker. For tighter subtitle
-    sync set ``subtitle_provider = "whisper"``.
-    """
-    text = (text or "").strip()
-    if not text:
-        logger.error("Chatterbox TTS text is empty")
-        return None
-
-    base_url = (config.chatterbox.get("base_url", "") or "").strip().rstrip("/")
-    if not base_url:
-        logger.error(
-            "Chatterbox base_url is not set, please configure [chatterbox] base_url in config.toml"
-        )
-        return None
-
-    api_key = config.chatterbox.get("api_key", "")
-    if not model_id:
-        model_id = config.chatterbox.get("model_id", "chatterbox") or "chatterbox"
-
-    url = f"{base_url}/audio/speech"
-    headers = {"Content-Type": "application/json"}
-    if api_key:
-        headers["Authorization"] = f"Bearer {api_key}"
-    payload = {
-        "model": model_id,
-        "input": text,
-        "voice": voice,
-        "response_format": "mp3",
-        # OpenAI speech API accepts speed 0.25-4.0; MoneyPrinterTurbo's rate is a
-        # 1.0-centred multiplier, so it maps directly (clamped to the valid range).
-        "speed": max(0.25, min(4.0, float(voice_rate or 1.0))),
-    }
-    # voice_volume is accepted for parity with the other TTS providers but is
-    # intentionally not sent: the OpenAI /audio/speech contract has no volume
-    # field, so Chatterbox servers ignore it. Adjust loudness via voice_rate
-    # (speed) or in post-processing instead.
-
-    for i in range(3):
-        try:
-            logger.info(f"start chatterbox tts, voice: {voice}, try: {i + 1}")
-            ensure_file_path_exists(voice_file)
-
-            response = requests.post(url, json=payload, headers=headers, timeout=120)
-            if response.status_code != 200:
-                logger.error(
-                    f"chatterbox tts failed with status {response.status_code}: {response.text[:200]}"
-                )
-                continue
-
-            with open(voice_file, "wb") as f:
-                f.write(response.content)
-
-            audio_clip = AudioFileClip(voice_file)
-            try:
-                audio_duration = audio_clip.duration
-            finally:
-                audio_clip.close()
-
-            sub_maker = ensure_legacy_submaker_fields(SubMaker())
-            logger.success(f"chatterbox tts succeeded: {voice_file}")
-            return populate_legacy_submaker_with_full_text(
-                sub_maker=sub_maker,
-                text=text,
-                audio_duration_seconds=audio_duration,
-            )
-        except Exception as e:
-            logger.error(f"chatterbox tts failed: {str(e)}")
-
-    return None
-
-
-# Fish Audio supported models.
-FISH_AUDIO_MODELS = ("s2.1-pro-free", "s2.1-pro", "s2-pro")
-FISH_AUDIO_DEFAULT_MODEL = "s2.1-pro-free"
-
-
-def fish_audio_tts(
-    text: str,
-    voice_file: str,
-    voice_rate: float = 1.0,
-    voice_volume: float = 1.0,
-    reference_id: str | None = None,
-) -> Union[SubMaker, None]:
-    """Generate speech using Fish Audio TTS API.
-
-    The model is read from ``config.fish_audio["model"]`` (single source of
-    truth).  ``reference_id`` selects a public or cloned voice; when *None*
-    Fish Audio's built-in default voice is used.
-
-    ``voice_rate`` is mapped to the ``prosody.speed`` field (0.5–2.0) and
-    ``voice_volume`` is converted from a linear multiplier to dB for the
-    ``prosody.volume`` field (-20.0–20.0 dB).
-    """
-    text = (text or "").strip()
-    if not text:
-        logger.error("Fish Audio TTS text is empty")
-        return None
-
-    api_key = get_fish_audio_api_key()
-    if not api_key:
-        logger.error(
-            "Fish Audio API key is not set. Please set it in config.toml "
-            "[fish_audio] or FISH_API_KEY environment variable."
-        )
-        return None
-
-    model_name = str(
-        config.fish_audio.get("model", FISH_AUDIO_DEFAULT_MODEL)
-        or FISH_AUDIO_DEFAULT_MODEL
-    ).strip()
-    if model_name not in FISH_AUDIO_MODELS:
-        logger.warning(
-            f"Unknown Fish Audio model '{model_name}', falling back to "
-            f"'{FISH_AUDIO_DEFAULT_MODEL}'"
-        )
-        model_name = FISH_AUDIO_DEFAULT_MODEL
-
-    # Map voice_rate → prosody.speed (0.5–2.0)
-    try:
-        speed = max(0.5, min(2.0, float(voice_rate or 1.0)))
-    except (TypeError, ValueError):
-        speed = 1.0
-
-    # Map voice_volume (linear multiplier) → prosody.volume (dB, -20–20).
-    # A multiplier of 1.0 → 0 dB; 0.1 → -20 dB; 2.0 → +6 dB.
-    import math
-    try:
-        vol = float(voice_volume or 1.0)
-        if vol <= 0:
-            volume_db = -20.0
-        else:
-            volume_db = max(-20.0, min(20.0, 20.0 * math.log10(vol)))
-    except (TypeError, ValueError):
-        volume_db = 0.0
-
-    url = "https://api.fish.audio/v1/tts"
-    headers = {
-        "Authorization": f"Bearer {api_key}",
-        "Content-Type": "application/json",
-        "model": model_name,
-    }
-    payload: dict = {
-        "text": text,
-        "format": "mp3",
-        "prosody": {
-            "speed": speed,
-            "volume": volume_db,
-        },
-    }
-    if reference_id:
-        payload["reference_id"] = reference_id
-
-    for i in range(3):
-        try:
-            logger.info(
-                f"start fish audio tts, model: {model_name}, "
-                f"ref: {reference_id or 'default'}, try: {i + 1}"
-            )
-            ensure_file_path_exists(voice_file)
-
-            response = requests.post(url, json=payload, headers=headers, timeout=60)
-            if response.status_code == 401:
-                logger.error(
-                    "Fish Audio TTS failed: Invalid API key (401). "
-                    "Check config.toml [fish_audio] api_key or FISH_API_KEY."
-                )
-                return None
-            if response.status_code == 402:
-                logger.error(
-                    "Fish Audio TTS failed: Insufficient API credit (402). "
-                    "Please check your account balance at "
-                    "https://fish.audio/app/developers or verify your model and billing tier."
-                )
-                return None
-            if response.status_code == 429:
-                logger.warning(
-                    "Fish Audio TTS rate limited (429), retrying..."
-                )
-                continue
-            if response.status_code != 200:
-                logger.error(
-                    f"fish audio tts failed with status "
-                    f"{response.status_code}: {response.text[:200]}"
-                )
-                continue
-
-            # Validate response contains audio data
-            if not response.content or len(response.content) < 100:
-                logger.error(
-                    "Fish Audio TTS returned empty or invalid audio data"
-                )
-                continue
-
-            with open(voice_file, "wb") as f:
-                f.write(response.content)
-
-            audio_clip = AudioFileClip(voice_file)
-            try:
-                audio_duration = audio_clip.duration
-            finally:
-                audio_clip.close()
-
-            sub_maker = ensure_legacy_submaker_fields(SubMaker())
-            logger.success(f"fish audio tts succeeded: {voice_file}")
-            return populate_legacy_submaker_with_full_text(
-                sub_maker=sub_maker,
-                text=text,
-                audio_duration_seconds=audio_duration,
-            )
-        except Exception as e:
-            logger.error(f"fish audio tts failed: {str(e)}")
-
-    return None
-
-
 def _format_text(text: str) -> str:
     """
     清理字幕对齐前的脚本文本。
@@ -1925,7 +356,7 @@ def _build_subtitle_formatter():
 
 # 阿拉伯语变音符号和 Tatweel 拉长符在 edge-tts 返回文本中可能出现，
 # 这些字符不影响语义，但会导致脚本文本和字幕 cue 字符串精确匹配失败。
-_ARABIC_DIACRITICS = re.compile("[\u0610-\u061A\u064B-\u065F\u0670\u0640\u06D6-\u06ED]")
+_ARABIC_DIACRITICS = re.compile("[ؐ-ًؚ-ٰٟـۖ-ۭ]")
 
 
 def _normalize_arabic(text: str) -> str:
@@ -2203,33 +634,6 @@ if __name__ == "__main__":
             "zh-CN-YunyangNeural",
             "zh-CN-YunxiNeural",
         ]
-        text = """
-        静夜思是唐代诗人李白创作的一首五言古诗。这首诗描绘了诗人在寂静的夜晚，看到窗前的明月，不禁想起远方的家乡和亲人，表达了他对家乡和亲人的深深思念之情。全诗内容是：“床前明月光，疑是地上霜。举头望明月，低头思故乡。”在这短短的四句诗中，诗人通过“明月”和“思故乡”的意象，巧妙地表达了离乡背井人的孤独与哀愁。首句“床前明月光”设景立意，通过明亮的月光引出诗人的遐想；“疑是地上霜”增添了夜晚的寒冷感，加深了诗人的孤寂之情；“举头望明月”和“低头思故乡”则是情感的升华，展现了诗人内心深处的乡愁和对家的渴望。这首诗简洁明快，情感真挚，是中国古典诗歌中非常著名的一首，也深受后人喜爱和推崇。
-            """
-
-        text = """
-        What is the meaning of life? This question has puzzled philosophers, scientists, and thinkers of all kinds for centuries. Throughout history, various cultures and individuals have come up with their interpretations and beliefs around the purpose of life. Some say it's to seek happiness and self-fulfillment, while others believe it's about contributing to the welfare of others and making a positive impact in the world. Despite the myriad of perspectives, one thing remains clear: the meaning of life is a deeply personal concept that varies from one person to another. It's an existential inquiry that encourages us to reflect on our values, desires, and the essence of our existence.
-        """
-
-        text = """
-               预计未来3天深圳冷空气活动频繁，未来两天持续阴天有小雨，出门带好雨具；
-               10-11日持续阴天有小雨，日温差小，气温在13-17℃之间，体感阴凉；
-               12日天气短暂好转，早晚清凉；
-                   """
-
-        text = "[Opening scene: A sunny day in a suburban neighborhood. A young boy named Alex, around 8 years old, is playing in his front yard with his loyal dog, Buddy.]\n\n[Camera zooms in on Alex as he throws a ball for Buddy to fetch. Buddy excitedly runs after it and brings it back to Alex.]\n\nAlex: Good boy, Buddy! You're the best dog ever!\n\n[Buddy barks happily and wags his tail.]\n\n[As Alex and Buddy continue playing, a series of potential dangers loom nearby, such as a stray dog approaching, a ball rolling towards the street, and a suspicious-looking stranger walking by.]\n\nAlex: Uh oh, Buddy, look out!\n\n[Buddy senses the danger and immediately springs into action. He barks loudly at the stray dog, scaring it away. Then, he rushes to retrieve the ball before it reaches the street and gently nudges it back towards Alex. Finally, he stands protectively between Alex and the stranger, growling softly to warn them away.]\n\nAlex: Wow, Buddy, you're like my superhero!\n\n[Just as Alex and Buddy are about to head inside, they hear a loud crash from a nearby construction site. They rush over to investigate and find a pile of rubble blocking the path of a kitten trapped underneath.]\n\nAlex: Oh no, Buddy, we have to help!\n\n[Buddy barks in agreement and together they work to carefully move the rubble aside, allowing the kitten to escape unharmed. The kitten gratefully nuzzles against Buddy, who responds with a friendly lick.]\n\nAlex: We did it, Buddy! We saved the day again!\n\n[As Alex and Buddy walk home together, the sun begins to set, casting a warm glow over the neighborhood.]\n\nAlex: Thanks for always being there to watch over me, Buddy. You're not just my dog, you're my best friend.\n\n[Buddy barks happily and nuzzles against Alex as they disappear into the sunset, ready to face whatever adventures tomorrow may bring.]\n\n[End scene.]"
-
-        text = "大家好，我是乔哥，一个想帮你把信用卡全部还清的家伙！\n今天我们要聊的是信用卡的取现功能。\n你是不是也曾经因为一时的资金紧张，而拿着信用卡到ATM机取现？如果是，那你得好好看看这个视频了。\n现在都2024年了，我以为现在不会再有人用信用卡取现功能了。前几天一个粉丝发来一张图片，取现1万。\n信用卡取现有三个弊端。\n一，信用卡取现功能代价可不小。会先收取一个取现手续费，比如这个粉丝，取现1万，按2.5%收取手续费，收取了250元。\n二，信用卡正常消费有最长56天的免息期，但取现不享受免息期。从取现那一天开始，每天按照万5收取利息，这个粉丝用了11天，收取了55元利息。\n三，频繁的取现行为，银行会认为你资金紧张，会被标记为高风险用户，影响你的综合评分和额度。\n那么，如果你资金紧张了，该怎么办呢？\n乔哥给你支一招，用破思机摩擦信用卡，只需要少量的手续费，而且还可以享受最长56天的免息期。\n最后，如果你对玩卡感兴趣，可以找乔哥领取一本《卡神秘籍》，用卡过程中遇到任何疑惑，也欢迎找乔哥交流。\n别忘了，关注乔哥，回复用卡技巧，免费领取《2024用卡技巧》，让我们一起成为用卡高手！"
-
-        text = """
-        2023全年业绩速览
-公司全年累计实现营业收入1476.94亿元，同比增长19.01%，归母净利润747.34亿元，同比增长19.16%。EPS达到59.49元。第四季度单季，营业收入444.25亿元，同比增长20.26%，环比增长31.86%；归母净利润218.58亿元，同比增长19.33%，环比增长29.37%。这一阶段
-的业绩表现不仅突显了公司的增长动力和盈利能力，也反映出公司在竞争激烈的市场环境中保持了良好的发展势头。
-2023年Q4业绩速览
-第四季度，营业收入贡献主要增长点；销售费用高增致盈利能力承压；税金同比上升27%，扰动净利率表现。
-业绩解读
-利润方面，2023全年贵州茅台，>归母净利润增速为19%，其中营业收入正贡献18%，营业成本正贡献百分之一，管理费用正贡献百分之一点四。(注：归母净利润增速值=营业收入增速+各科目贡献，展示贡献/拖累的前四名科目，且要求贡献值/净利润增速>15%)
-"""
         text = "静夜思是唐代诗人李白创作的一首五言古诗。这首诗描绘了诗人在寂静的夜晚，看到窗前的明月，不禁想起远方的家乡和亲人"
 
         text = _format_text(text)
