+++ app/services/voice_azure_v2.py
@@ -1,0 +1,131 @@
+"""Azure Speech SDK provider (``azure_tts_v2``, the ``*-V2`` voice suffix)."""
+
+from datetime import datetime
+from typing import Union
+from xml.sax.saxutils import escape
+
+from edge_tts import SubMaker
+from loguru import logger
+
+from app.config import config
+from app.services.voice_common import ensure_legacy_submaker_fields, is_azure_v2_voice
+
+
+def _build_azure_v2_ssml(text: str, voice_name: str, voice_rate: float) -> str:
+    """构造 Azure Speech V2 使用的 SSML，并安全规范化语速参数。"""
+    try:
+        normalized_rate = float(voice_rate)
+    except (TypeError, ValueError):
+        normalized_rate = 1.0
+    normalized_rate = max(0.25, min(4.0, normalized_rate))
+
+    voice_locale_parts = voice_name.split("-", 2)
+    voice_locale = (
+        "-".join(voice_locale_parts[:2])
+        if len(voice_locale_parts) >= 2
+        else "en-US"
+    )
+    escaped_text = escape(text)
+    escaped_voice_name = escape(voice_name, {'"': "&quot;"})
+    return (
+        '<speak version="1.0" xmlns="http://www.w3.org/2001/10/synthesis" '
+        f'xml:lang="{voice_locale}">'
+        f'<voice name="{escaped_voice_name}">'
+        f'<prosody rate="{normalized_rate:g}">{escaped_text}</prosody>'
+        "</voice></speak>"
+    )
+
+
+def azure_tts_v2(
+    text: str,
+    voice_name: str,
+    voice_file: str,
+    voice_rate: float = 1.0,
+) -> Union[SubMaker, None]:
+    voice_name = is_azure_v2_voice(voice_name)
+    if not voice_name:
+        logger.error(f"invalid voice name: {voice_name}")
+        raise ValueError(f"invalid voice name: {voice_name}")
+    text = text.strip()
+    ssml = _build_azure_v2_ssml(text, voice_name, voice_rate)
+
+    def _format_duration_to_offset(duration) -> int:
+        if isinstance(duration, str):
+            time_obj = datetime.strptime(duration, "%H:%M:%S.%f")
+            milliseconds = (
+                (time_obj.hour * 3600000)
+                + (time_obj.minute * 60000)
+                + (time_obj.second * 1000)
+                + (time_obj.microsecond // 1000)
+            )
+            return milliseconds * 10000
+
+        if isinstance(duration, int):
+            return duration
+
+        return 0
+
+    for i in range(3):
+        try:
+            logger.info(
+                f"start, voice name: {voice_name}, rate: {voice_rate}, try: {i + 1}"
+            )
+
+            import azure.cognitiveservices.speech as speechsdk
+
+            sub_maker = ensure_legacy_submaker_fields(SubMaker())
+
+            def speech_synthesizer_word_boundary_cb(evt: speechsdk.SessionEventArgs):
+                duration = _format_duration_to_offset(str(evt.duration))
+                offset = _format_duration_to_offset(evt.audio_offset)
+                sub_maker.subs.append(evt.text)
+                sub_maker.offset.append((offset, offset + duration))
+
+            # Creates an instance of a speech config with specified subscription key and region.
+            speech_key = config.azure.get("speech_key", "")
+            service_region = config.azure.get("speech_region", "")
+            if not speech_key or not service_region:
+                logger.error("Azure speech key or region is not set")
+                return None
+
+            audio_config = speechsdk.audio.AudioOutputConfig(
+                filename=voice_file, use_default_speaker=True
+            )
+            speech_config = speechsdk.SpeechConfig(
+                subscription=speech_key, region=service_region
+            )
+            speech_config.speech_synthesis_voice_name = voice_name
+            speech_config.set_property(
+                property_id=speechsdk.PropertyId.SpeechServiceResponse_RequestWordBoundary,
+                value="true",
+            )
+
+            speech_config.set_speech_synthesis_output_format(
+                speechsdk.SpeechSynthesisOutputFormat.Audio48Khz192KBitRateMonoMp3
+            )
+            speech_synthesizer = speechsdk.SpeechSynthesizer(
+                audio_config=audio_config, speech_config=speech_config
+            )
+            speech_synthesizer.synthesis_word_boundary.connect(
+                speech_synthesizer_word_boundary_cb
+            )
+
+            # speak_text_async() 不支持语速参数。使用 SSML prosody 后，试听和
+            # 正式生成都会按 WebUI/API 传入的 voice_rate 调整语速。
+            result = speech_synthesizer.speak_ssml_async(ssml).get()
+            if result.reason == speechsdk.ResultReason.SynthesizingAudioCompleted:
+                logger.success(f"azure v2 speech synthesis succeeded: {voice_file}")
+                return sub_maker
+            elif result.reason == speechsdk.ResultReason.Canceled:
+                cancellation_details = result.cancellation_details
+                logger.error(
+                    f"azure v2 speech synthesis canceled: {cancellation_details.reason}"
+                )
+                if cancellation_details.reason == speechsdk.CancellationReason.Error:
+                    logger.error(
+                        f"azure v2 speech synthesis error: {cancellation_details.error_details}"
+                    )
+            logger.info(f"completed, output file: {voice_file}")
+        except Exception as e:
+            logger.error(f"failed, error: {str(e)}")
+    return None
