+++ app/services/voice_chatterbox.py
@@ -1,0 +1,131 @@
+"""Self-hosted Chatterbox TTS provider."""
+
+from typing import Union
+
+import requests
+from edge_tts import SubMaker
+from loguru import logger
+from moviepy.audio.io.AudioFileClip import AudioFileClip
+
+from app.config import config
+from app.services.voice_common import (
+    ensure_file_path_exists,
+    ensure_legacy_submaker_fields,
+    populate_legacy_submaker_with_full_text,
+)
+
+
+def get_chatterbox_voices() -> list[str]:
+    """Return the configured Chatterbox voices.
+
+    Chatterbox is self-hosted, so there is no global voice catalog. Operators
+    list the voice names exposed by their server via ``[chatterbox] voices``
+    (a TOML array, or a comma-separated string). Each entry is normalised to
+    the ``chatterbox:<name>`` format used by the TTS dispatcher.
+    """
+    voices = config.chatterbox.get("voices", []) or []
+    if isinstance(voices, str):
+        voices = [v.strip() for v in voices.split(",") if v.strip()]
+    result = []
+    for v in voices:
+        v = str(v).strip()
+        if not v:
+            continue
+        result.append(v if v.startswith("chatterbox:") else f"chatterbox:{v}")
+    if not result:
+        # keep the dropdown usable even before any voice is configured
+        result = ["chatterbox:default-Female"]
+    return result
+
+
+def is_chatterbox_voice(voice_name: str) -> bool:
+    return (voice_name or "").startswith("chatterbox:")
+
+
+def chatterbox_tts(
+    text: str,
+    voice: str,
+    voice_file: str,
+    voice_rate: float = 1.0,
+    voice_volume: float = 1.0,
+    model_id: str = "",
+) -> Union[SubMaker, None]:
+    """Generate speech with a self-hosted Chatterbox TTS server.
+
+    Chatterbox (Resemble AI, MIT) is an open-source, locally hosted TTS model
+    with zero-shot voice cloning — a self-hostable alternative to ElevenLabs.
+    This talks to an OpenAI-compatible ``/audio/speech`` endpoint, so it works
+    with the common community servers (e.g. devnen/Chatterbox-TTS-Server,
+    travisvn/chatterbox-tts-api). Configure ``[chatterbox] base_url`` (and an
+    optional ``api_key``).
+
+    Like ElevenLabs, Chatterbox does not return word-level timestamps, so the
+    subtitle path falls back to the full-text SubMaker. For tighter subtitle
+    sync set ``subtitle_provider = "whisper"``.
+    """
+    text = (text or "").strip()
+    if not text:
+        logger.error("Chatterbox TTS text is empty")
+        return None
+
+    base_url = (config.chatterbox.get("base_url", "") or "").strip().rstrip("/")
+    if not base_url:
+        logger.error(
+            "Chatterbox base_url is not set, please configure [chatterbox] base_url in config.toml"
+        )
+        return None
+
+    api_key = config.chatterbox.get("api_key", "")
+    if not model_id:
+        model_id = config.chatterbox.get("model_id", "chatterbox") or "chatterbox"
+
+    url = f"{base_url}/audio/speech"
+    headers = {"Content-Type": "application/json"}
+    if api_key:
+        headers["Authorization"] = f"Bearer {api_key}"
+    payload = {
+        "model": model_id,
+        "input": text,
+        "voice": voice,
+        "response_format": "mp3",
+        # OpenAI speech API accepts speed 0.25-4.0; MoneyPrinterTurbo's rate is a
+        # 1.0-centred multiplier, so it maps directly (clamped to the valid range).
+        "speed": max(0.25, min(4.0, float(voice_rate or 1.0))),
+    }
+    # voice_volume is accepted for parity with the other TTS providers but is
+    # intentionally not sent: the OpenAI /audio/speech contract has no volume
+    # field, so Chatterbox servers ignore it. Adjust loudness via voice_rate
+    # (speed) or in post-processing instead.
+
+    for i in range(3):
+        try:
+            logger.info(f"start chatterbox tts, voice: {voice}, try: {i + 1}")
+            ensure_file_path_exists(voice_file)
+
+            response = requests.post(url, json=payload, headers=headers, timeout=120)
+            if response.status_code != 200:
+                logger.error(
+                    f"chatterbox tts failed with status {response.status_code}: {response.text[:200]}"
+                )
+                continue
+
+            with open(voice_file, "wb") as f:
+                f.write(response.content)
+
+            audio_clip = AudioFileClip(voice_file)
+            try:
+                audio_duration = audio_clip.duration
+            finally:
+                audio_clip.close()
+
+            sub_maker = ensure_legacy_submaker_fields(SubMaker())
+            logger.success(f"chatterbox tts succeeded: {voice_file}")
+            return populate_legacy_submaker_with_full_text(
+                sub_maker=sub_maker,
+                text=text,
+                audio_duration_seconds=audio_duration,
+            )
+        except Exception as e:
+            logger.error(f"chatterbox tts failed: {str(e)}")
+
+    return None
