+++ app/services/voice_common.py
@@ -1,0 +1,169 @@
+"""Shared helpers used by every TTS provider module under ``app.services.voice_*``.
+
+Kept dependency-free from ``app.services.voice`` and from any provider module,
+so the import graph stays a clean DAG: this module has no knowledge of who
+calls it, ``voice.py`` and every ``voice_<provider>.py`` import from here, and
+nothing imports back from those into this module.
+"""
+
+import math
+import os
+
+from edge_tts import SubMaker
+
+from app.utils import utils
+
+
+def mktimestamp(time_unit: float) -> str:
+    """
+    将 edge_tts 使用的 100 纳秒时间单位转换为字幕时间戳。
+
+    edge_tts 7.x 不再导出旧版本里的 `mktimestamp`，但项目里旧字幕链路
+    还需要这个格式化函数来兼容 Azure v2、Gemini、SiliconFlow 这些
+    手工构造的字幕时间轴，因此这里内置一个等价实现。
+    """
+    hour = math.floor(time_unit / 10**7 / 3600)
+    minute = math.floor((time_unit / 10**7 / 60) % 60)
+    seconds = (time_unit / 10**7) % 60
+    return f"{hour:02d}:{minute:02d}:{seconds:06.3f}"
+
+
+def _configure_pydub_ffmpeg(audio_segment_cls):
+    configured_ffmpeg = utils.get_ffmpeg_binary()
+    if configured_ffmpeg:
+        audio_segment_cls.converter = configured_ffmpeg
+
+
+def parse_voice_name(name: str):
+    # zh-CN-XiaoyiNeural-Female
+    # zh-CN-YunxiNeural-Male
+    # zh-CN-XiaoxiaoMultilingualNeural-V2-Female
+    name = name.replace("-Female", "").replace("-Male", "").strip()
+    return name
+
+
+def is_azure_v2_voice(voice_name: str):
+    voice_name = parse_voice_name(voice_name)
+    if voice_name.endswith("-V2"):
+        return voice_name.replace("-V2", "").strip()
+    return ""
+
+
+def convert_rate_to_percent(rate: float) -> str:
+    # edge-tts requires a sign-prefixed percentage (e.g. "+0%", "-20%").
+    # Rounding can yield 0 for rates near but not equal to 1.0 (e.g. 1.004,
+    # 0.997); those must still be returned as "+0%", not the unsigned "0%"
+    # which edge-tts rejects with ValueError: Invalid rate '0%'.
+    # API 或批处理调用可能传入 0、0.0、None 或无法转换的空值；这些值不代表
+    # 合法语速，直接计算会变成 -100% 或抛异常。这里统一回退到正常语速，
+    # 避免生成极慢音频或让 TTS 流程在边界输入下失败。
+    try:
+        rate = float(rate)
+    except (TypeError, ValueError):
+        rate = 1.0
+    if rate <= 0:
+        rate = 1.0
+    percent = round((rate - 1.0) * 100)
+    if percent >= 0:
+        return f"+{percent}%"
+    return f"{percent}%"
+
+
+def ensure_file_path_exists(file_path: str) -> None:
+    """
+    确保输出文件所在目录一定存在。
+
+    这里单独做一层兜底，是因为 edge_tts 7.x 在真正发起网络请求之前，
+    就会先打开目标音频文件；如果目录不存在，会直接因为本地文件路径报错，
+    从而掩盖真正的 TTS 行为结果。
+    """
+    dir_path = os.path.dirname(file_path)
+    if dir_path:
+        os.makedirs(dir_path, exist_ok=True)
+
+
+def ensure_legacy_submaker_fields(sub_maker: SubMaker) -> SubMaker:
+    """
+    为项目里仍然沿用旧字幕结构的调用方补齐兼容字段。
+
+    edge_tts 7.x 的 `SubMaker` 主要暴露 `cues/get_srt()`，但项目里 Azure v2、
+    Gemini、SiliconFlow 这些路径仍然会直接读写 `subs/offset`。这里统一补齐，
+    避免升级 edge_tts 后这些非 edge 路径被连带破坏。
+    """
+    if not hasattr(sub_maker, "subs"):
+        sub_maker.subs = []
+    if not hasattr(sub_maker, "offset"):
+        sub_maker.offset = []
+    return sub_maker
+
+
+def populate_legacy_submaker_with_full_text(
+    sub_maker: SubMaker, text: str, audio_duration_seconds: float
+) -> SubMaker:
+    """
+    用整段文本填充项目历史沿用的 `subs/offset` 字幕结构。
+
+    背景：
+    1. edge_tts 7.x 的 `SubMaker` 不再提供旧版本里的 `create_sub()`；
+    2. 项目里 Gemini、SiliconFlow 等非 edge 路径依然需要返回一个
+       带 `subs/offset` 的对象，供后续统一计算音频时长和生成字幕；
+    3. 对于拿不到逐词边界的 TTS 服务，需要至少按脚本断句切成多个片段，
+       这样后续 `subtitle_provider=edge` 的聚合逻辑才能继续工作，而不是
+       因为整段文本无法和脚本断句逐行匹配而回退 Whisper。
+
+    Args:
+        sub_maker: 需要写入兼容字段的字幕对象
+        text: 原始脚本文本
+        audio_duration_seconds: 音频总时长，单位秒
+
+    Returns:
+        已填充兼容字幕数据的 SubMaker 对象
+    """
+    sub_maker = ensure_legacy_submaker_fields(sub_maker)
+
+    # 清空旧值，避免调用方重复复用对象时出现脏数据叠加。
+    sub_maker.subs = []
+    sub_maker.offset = []
+
+    normalized_text = (text or "").strip()
+    if not normalized_text:
+        return sub_maker
+
+    audio_duration_100ns = max(int(audio_duration_seconds * 10000000), 1)
+
+    # Gemini / SiliconFlow 这类路径拿不到逐词边界时，仍然尽量沿用项目
+    # 原来的“按标点断句 + 按字符数比例分配时长”的策略。这样既能让
+    # create_subtitle() 匹配脚本断句，也能避免再次回退 Whisper。
+    sentences = utils.split_string_by_punctuations(normalized_text)
+    if not sentences:
+        sentences = [normalized_text]
+
+    total_chars = sum(len(sentence) for sentence in sentences)
+    if total_chars <= 0:
+        sub_maker.subs.append(normalized_text)
+        sub_maker.offset.append((0, audio_duration_100ns))
+        return sub_maker
+
+    current_offset = 0
+    for index, sentence in enumerate(sentences):
+        cleaned_sentence = sentence.strip()
+        if not cleaned_sentence:
+            continue
+
+        # 前面的句子按字符数比例分配时长，最后一句兜底吃掉剩余时长，
+        # 避免整数取整导致总时长丢失或字幕结束时间短于音频。
+        if index == len(sentences) - 1:
+            sentence_end = audio_duration_100ns
+        else:
+            sentence_chars = len(cleaned_sentence)
+            sentence_duration = max(
+                int(audio_duration_100ns * (sentence_chars / total_chars)),
+                1,
+            )
+            sentence_end = min(current_offset + sentence_duration, audio_duration_100ns)
+
+        sub_maker.subs.append(cleaned_sentence)
+        sub_maker.offset.append((current_offset, sentence_end))
+        current_offset = sentence_end
+
+    return sub_maker
