+++ app/services/voice_elevenlabs.py
@@ -1,0 +1,148 @@
+"""ElevenLabs TTS provider."""
+
+import os
+from typing import Union
+
+import requests
+from edge_tts import SubMaker
+from loguru import logger
+from moviepy.audio.io.AudioFileClip import AudioFileClip
+
+from app.config import config
+from app.services.voice_common import (
+    ensure_file_path_exists,
+    ensure_legacy_submaker_fields,
+    populate_legacy_submaker_with_full_text,
+)
+
+
+def get_elevenlabs_voices(api_key: str) -> list[str]:
+    if not api_key:
+        return []
+    try:
+        url = "https://api.elevenlabs.io/v2/voices"
+        params = {"is_favorite": "true", "page_size": 100}
+        headers = {"xi-api-key": api_key}
+        response = requests.get(url, params=params, headers=headers, timeout=10)
+        if response.status_code != 200:
+            logger.warning(
+                f"ElevenLabs voices fetch failed with status {response.status_code}: {response.text}"
+            )
+            return []
+        data = response.json()
+        voices = data.get("voices", [])
+        return [
+            f"elevenlabs:{v['voice_id']}:{v['name']}"
+            for v in voices
+            if v.get("voice_id") and v.get("name") and v.get("status") != "disabled"
+        ]
+    except Exception as e:
+        logger.warning(f"ElevenLabs voices fetch failed: {str(e)}")
+        return []
+
+
+def is_elevenlabs_voice(voice_name: str) -> bool:
+    return (voice_name or "").startswith("elevenlabs:")
+
+
+def get_elevenlabs_api_key() -> str:
+    """
+    读取 ElevenLabs TTS 使用的 API Key。
+
+    配置文件优先，环境变量仅作为未配置时的后备来源。WebUI 和配乐功能已经
+    支持 ``ELEVENLABS_API_KEY``，TTS 必须使用相同规则，否则仅通过容器环境
+    变量部署时，音色列表可正常加载，真正合成语音却会误报未配置 Key。
+    """
+    configured_key = str(config.elevenlabs.get("api_key", "") or "").strip()
+    return configured_key or os.getenv("ELEVENLABS_API_KEY", "").strip()
+
+
+def elevenlabs_tts(
+    text: str,
+    voice_id: str,
+    voice_file: str,
+    voice_rate: float = 1.0,
+    voice_volume: float = 1.0,
+    model_id: str = "",
+) -> Union[SubMaker, None]:
+    text = (text or "").strip()
+    if not text:
+        logger.error("ElevenLabs TTS text is empty")
+        return None
+
+    api_key = get_elevenlabs_api_key()
+    if not api_key:
+        logger.error("ElevenLabs API key is not set")
+        return None
+
+    if not model_id:
+        model_id = config.elevenlabs.get("model_id", "eleven_multilingual_v2")
+
+    url = f"https://api.elevenlabs.io/v1/text-to-speech/{voice_id}"
+    headers = {
+        "xi-api-key": api_key,
+        "Content-Type": "application/json",
+    }
+    payload = {
+        "text": text,
+        "model_id": model_id,
+        "voice_settings": {
+            "stability": 0.5,
+            "similarity_boost": 0.75,
+            "style": 0.0,
+            "use_speaker_boost": True,
+        },
+    }
+
+    # Errors where retrying will never help (auth/access/validation failures).
+    _NON_RETRYABLE_CODES = {401, 403, 422}
+    _NON_RETRYABLE_STATUSES = {"voice_disabled", "voice_access_denied", "unauthorized"}
+
+    for i in range(3):
+        try:
+            logger.info(f"start elevenlabs tts, voice_id: {voice_id}, try: {i + 1}")
+            ensure_file_path_exists(voice_file)
+
+            response = requests.post(url, json=payload, headers=headers, timeout=60)
+            if response.status_code != 200:
+                error_status = ""
+                try:
+                    detail = response.json().get("detail", {})
+                    if isinstance(detail, dict):
+                        error_status = detail.get("status", "")
+                except Exception:
+                    pass
+
+                if response.status_code in _NON_RETRYABLE_CODES or error_status in _NON_RETRYABLE_STATUSES:
+                    logger.error(
+                        f"ElevenLabs TTS failed (non-retryable) — voice_id: {voice_id}, "
+                        f"status: {response.status_code}, error: {error_status or response.text[:200]}. "
+                        "Please select a different ElevenLabs voice."
+                    )
+                    return None
+
+                logger.error(
+                    f"elevenlabs tts failed with status {response.status_code}: {response.text[:200]}"
+                )
+                continue
+
+            with open(voice_file, "wb") as f:
+                f.write(response.content)
+
+            audio_clip = AudioFileClip(voice_file)
+            try:
+                audio_duration = audio_clip.duration
+            finally:
+                audio_clip.close()
+
+            sub_maker = ensure_legacy_submaker_fields(SubMaker())
+            logger.success(f"elevenlabs tts succeeded: {voice_file}")
+            return populate_legacy_submaker_with_full_text(
+                sub_maker=sub_maker,
+                text=text,
+                audio_duration_seconds=audio_duration,
+            )
+        except Exception as e:
+            logger.error(f"elevenlabs tts failed: {str(e)}")
+
+    return None
