+++ app/services/voice_fish_audio.py
@@ -1,0 +1,200 @@
+"""Fish Audio TTS provider."""
+
+import math
+import os
+from typing import Union
+
+import requests
+from edge_tts import SubMaker
+from loguru import logger
+from moviepy.audio.io.AudioFileClip import AudioFileClip
+
+from app.config import config
+from app.services.voice_common import (
+    ensure_file_path_exists,
+    ensure_legacy_submaker_fields,
+    populate_legacy_submaker_with_full_text,
+)
+
+# Fish Audio supported models.
+FISH_AUDIO_MODELS = ("s2.1-pro-free", "s2.1-pro", "s2-pro")
+FISH_AUDIO_DEFAULT_MODEL = "s2.1-pro-free"
+
+
+def get_fish_audio_voices() -> list[str]:
+    """Return configured Fish Audio voices.
+
+    Each entry follows the format ``fish_audio:<reference_id>:<display_name>``.
+    When ``reference_id`` is "default", Fish Audio's built-in default voice is
+    used (no ``reference_id`` is sent in the API request).  Operators can list
+    additional public or cloned voices via ``[fish_audio] voices`` in the
+    config file.
+    """
+    result = [
+        "fish_audio:2324c907b9a94c64ab4afb941e5b3408:Clear Female-Female",
+        "fish_audio:7b6131ba75ba47c98a46c847db729ab6:Clear Male-Male",
+        "fish_audio:default:Default Voice",
+    ]
+    voices = config.fish_audio.get("voices", []) or []
+    if isinstance(voices, str):
+        voices = [v.strip() for v in voices.split(",") if v.strip()]
+    for entry in voices:
+        entry = str(entry).strip()
+        if not entry:
+            continue
+        if entry.startswith("fish_audio:"):
+            result.append(entry)
+        elif ":" in entry:
+            # "<reference_id>:<display_name>"
+            result.append(f"fish_audio:{entry}")
+        else:
+            # bare reference_id
+            result.append(f"fish_audio:{entry}:{entry}")
+    return result
+
+
+def is_fish_audio_voice(voice_name: str) -> bool:
+    return (voice_name or "").startswith("fish_audio:")
+
+
+def get_fish_audio_api_key() -> str:
+    configured_key = str(config.fish_audio.get("api_key", "") if hasattr(config, "fish_audio") and isinstance(config.fish_audio, dict) else "").strip()
+    return configured_key or os.getenv("FISH_API_KEY", "").strip()
+
+
+def fish_audio_tts(
+    text: str,
+    voice_file: str,
+    voice_rate: float = 1.0,
+    voice_volume: float = 1.0,
+    reference_id: str | None = None,
+) -> Union[SubMaker, None]:
+    """Generate speech using Fish Audio TTS API.
+
+    The model is read from ``config.fish_audio["model"]`` (single source of
+    truth).  ``reference_id`` selects a public or cloned voice; when *None*
+    Fish Audio's built-in default voice is used.
+
+    ``voice_rate`` is mapped to the ``prosody.speed`` field (0.5–2.0) and
+    ``voice_volume`` is converted from a linear multiplier to dB for the
+    ``prosody.volume`` field (-20.0–20.0 dB).
+    """
+    text = (text or "").strip()
+    if not text:
+        logger.error("Fish Audio TTS text is empty")
+        return None
+
+    api_key = get_fish_audio_api_key()
+    if not api_key:
+        logger.error(
+            "Fish Audio API key is not set. Please set it in config.toml "
+            "[fish_audio] or FISH_API_KEY environment variable."
+        )
+        return None
+
+    model_name = str(
+        config.fish_audio.get("model", FISH_AUDIO_DEFAULT_MODEL)
+        or FISH_AUDIO_DEFAULT_MODEL
+    ).strip()
+    if model_name not in FISH_AUDIO_MODELS:
+        logger.warning(
+            f"Unknown Fish Audio model '{model_name}', falling back to "
+            f"'{FISH_AUDIO_DEFAULT_MODEL}'"
+        )
+        model_name = FISH_AUDIO_DEFAULT_MODEL
+
+    # Map voice_rate → prosody.speed (0.5–2.0)
+    try:
+        speed = max(0.5, min(2.0, float(voice_rate or 1.0)))
+    except (TypeError, ValueError):
+        speed = 1.0
+
+    # Map voice_volume (linear multiplier) → prosody.volume (dB, -20–20).
+    # A multiplier of 1.0 → 0 dB; 0.1 → -20 dB; 2.0 → +6 dB.
+    try:
+        vol = float(voice_volume or 1.0)
+        if vol <= 0:
+            volume_db = -20.0
+        else:
+            volume_db = max(-20.0, min(20.0, 20.0 * math.log10(vol)))
+    except (TypeError, ValueError):
+        volume_db = 0.0
+
+    url = "https://api.fish.audio/v1/tts"
+    headers = {
+        "Authorization": f"Bearer {api_key}",
+        "Content-Type": "application/json",
+        "model": model_name,
+    }
+    payload: dict = {
+        "text": text,
+        "format": "mp3",
+        "prosody": {
+            "speed": speed,
+            "volume": volume_db,
+        },
+    }
+    if reference_id:
+        payload["reference_id"] = reference_id
+
+    for i in range(3):
+        try:
+            logger.info(
+                f"start fish audio tts, model: {model_name}, "
+                f"ref: {reference_id or 'default'}, try: {i + 1}"
+            )
+            ensure_file_path_exists(voice_file)
+
+            response = requests.post(url, json=payload, headers=headers, timeout=60)
+            if response.status_code == 401:
+                logger.error(
+                    "Fish Audio TTS failed: Invalid API key (401). "
+                    "Check config.toml [fish_audio] api_key or FISH_API_KEY."
+                )
+                return None
+            if response.status_code == 402:
+                logger.error(
+                    "Fish Audio TTS failed: Insufficient API credit (402). "
+                    "Please check your account balance at "
+                    "https://fish.audio/app/developers or verify your model and billing tier."
+                )
+                return None
+            if response.status_code == 429:
+                logger.warning(
+                    "Fish Audio TTS rate limited (429), retrying..."
+                )
+                continue
+            if response.status_code != 200:
+                logger.error(
+                    f"fish audio tts failed with status "
+                    f"{response.status_code}: {response.text[:200]}"
+                )
+                continue
+
+            # Validate response contains audio data
+            if not response.content or len(response.content) < 100:
+                logger.error(
+                    "Fish Audio TTS returned empty or invalid audio data"
+                )
+                continue
+
+            with open(voice_file, "wb") as f:
+                f.write(response.content)
+
+            audio_clip = AudioFileClip(voice_file)
+            try:
+                audio_duration = audio_clip.duration
+            finally:
+                audio_clip.close()
+
+            sub_maker = ensure_legacy_submaker_fields(SubMaker())
+            logger.success(f"fish audio tts succeeded: {voice_file}")
+            return populate_legacy_submaker_with_full_text(
+                sub_maker=sub_maker,
+                text=text,
+                audio_duration_seconds=audio_duration,
+            )
+        except Exception as e:
+            logger.error(f"fish audio tts failed: {str(e)}")
+
+    return None
